% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cumulative_volume.R
\name{add_cumulative_volume}
\alias{add_cumulative_volume}
\title{Add daily cumulative volumetric flows on an annual basis}
\usage{
add_cumulative_volume(data = NULL, dates = Date, values = Value,
  grouping = STATION_NUMBER, water_year = FALSE, water_year_start = 10)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groupings (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{grouping}{Column in the \code{data} data frame that contains unique identifiers for different data sets. 
Only required if using the data frame option of \code{data} and grouping column is not named 'STATION_NUMBER'.
Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column or identify 
another non-existing column name to remove this grouping. Identify another column if desired. Default \code{STATION_NUMBER}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}
}
\value{
A data frame of the source data with an additional column:
  \item{Cumul_Volume_m3}{cumulative volumetric flows for each day for each year, in units of cubic metres}
}
\description{
Add a column of daily cumulative volumetric flows on an annual basis to a streamflow dataset. Adds the volumetric discharge 
   from each day with the previous day(s) for each year, in units of cubic metres. The cumulative flows restart every year.
}
\examples{
\dontrun{

add_cumulative_volume(data = flow_data)

add_cumulative_volume(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
