% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cumulative_yield.R
\name{add_cumulative_yield}
\alias{add_cumulative_yield}
\title{Add daily cumulative runoff yield flows on an annual basis}
\usage{
add_cumulative_yield(flow_data = NULL, flow_dates = Date,
  flow_values = Value, flow_basin_areas = flow_basin_areas, HYDAT = NULL,
  basin_area = NA, water_year = FALSE, water_year_start = 10)
}
\arguments{
\item{flow_data}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{flow_dates}{A column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Default \code{Date}.}

\item{flow_values}{A column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Default \code{Value}.}

\item{flow_basin_areas}{A column in flow_data of upstream drainage basin areas used to calculate the daily yield. If not provided 
this function will create one based on \code{basin_area} values.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flow_data} argument is used.}

\item{basin_area}{Numeric. If no \code{flow_basin_area} provided in flow_data, used to determine basin areas. Leave blank if \code{HYDAT}
is used or a column in \code{flow_data} called 'STATION_NUMBER' contains a WSC station number, as the basin area will be extracted from HYDAT. 
Using \code{basin_area} will replace the HYDAT basin area. If setting basin areas for multiple stations without HYDAT, set them using 
\code{basin_area = c("08NM116" = 795, "08NM242" = 10)}; stations not listed will result in NA basin areas.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}
}
\value{
A data frame of the original flow_data or HYDAT data with an additional column:
  \item{Basin_Area_sqkm}{(if doesn't exist in data) area of upstream drainage basin area, in square kilometres}
  \item{Cumul_Yield_mm}{cumulative yield flows for each day for each year, in units of millimetres}
}
\description{
Add a column of daily cumulative runoff yield flows on an annual basis to a streamflow dataset. Adds the runoff yield 
   discharge from each day with the previous day(s) for each year, in units of millimetres. The cumulative flows restart every year.
   Converts cumulative discharge to a depth of water based on the upstream drainge basin area.
}
\examples{
\dontrun{

add_cumulative_yield(flow_data = flow_data, basin_area = 105.6)

add_cumulative_yield(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
