% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_daily_yield.R
\name{add_daily_yield}
\alias{add_daily_yield}
\title{Add daily volumetric runoff yields}
\usage{
add_daily_yield(flow_data = NULL, flow_values = Value,
  flow_basin_areas = Basin_Area_sqkm, HYDAT = NULL, basin_area = NA)
}
\arguments{
\item{flow_data}{Data frame. A data frame of daily mean flow data. Not required if \code{HYDAT} argument is used.}

\item{flow_values}{A column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Default \code{Value}.}

\item{flow_basin_areas}{A column in flow_data of upstream drainage basin areas used to calculate the daily yield. If not provided 
this function will create one based on \code{basin_area} values.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flow_data} argument is used.}

\item{basin_area}{Numeric. If no \code{flow_basin_area} provided in flow_data, used to determine basin areas. Leave blank if \code{HYDAT}
is used or a column in \code{flow_data} called 'STATION_NUMBER' contains a WSC station number, as the basin area will be extracted from HYDAT. 
Using \code{basin_area} will replace the HYDAT basin area. If setting basin areas for multiple stations without HYDAT, set them using 
\code{basin_area = c("08NM116" = 795, "08NM242" = 10)}; stations not listed will result in NA basin areas.}
}
\value{
A data frame of the original flow_data or HYDAT data with an additional columns:
  \item{Basin_Area_sqkm}{(if doesn't exist in data) area of upstream drainage basin area, in square kilometres}
  \item{Yield_mm}{daily runoff yield flow, in units of millimetres}
}
\description{
Add a column of daily runoff yields to a streamflow dataset, in units of millimetres. Converts the discharge to a depth
  of water based on the upstream drainge basin area.
}
\examples{
\dontrun{

add_daily_yield(flow_data = flow_data, basin_area = 104.5)

add_daily_yield(HYDAT = "08NM116")

}
}
