% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_variables.R
\name{add_date_variables}
\alias{add_date_variables}
\title{Add year, month, and day of year variables}
\usage{
add_date_variables(data = NULL, dates = Date, station_number = NULL,
  water_year = FALSE, water_year_start = 10)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the calendar year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}
}
\value{
A tibble data frame of the source data with additional columns:
  \item{Year}{calendar year}
  \item{Month}{numeric month (1 to 12)}
  \item{MonthName}{month name (Jan-Dec)}
  \item{DayofYear}{day of the year (1-365 or 366)}
  \item{WaterYear}{(optional) water year, designated by the calendar year in which it ends}
  \item{WaterDayofYear}{(optional) day of the water year (1-365 or 366), starting in the first month of the water year}
}
\examples{
\dontrun{

add_date_variables(data = flow_data)

add_date_variables(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
