% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rolling_means.R
\name{add_rolling_means}
\alias{add_rolling_means}
\title{Add rolling n-day averages}
\usage{
add_rolling_means(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, days = c(3, 7, 30), align = "right")
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{groups}{Column in the \code{data} data frame that contains unique identifiers for different data sets. 
Only required if using the data frame option of \code{data} and groups column is not named 'STATION_NUMBER'.
Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column or identify 
another non-existing column name to remove this grouping. Identify another column if desired. Default \code{STATION_NUMBER}.}

\item{days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}
}
\value{
A data frame of the source data with an additional column(s):
  \item{QnDay}{rolling means of the n-day flow values of the designated date and adjacent dates, direction of mean specified by align}
  Default additional columns:
  \item{Q3Day}{rolling means of the 3-day flow values of the designated date and previous 2 days (align="right")}
  \item{Q7Day}{rolling means of the 7-day flow values of the designated date and previous 6 days (align="right")}
  \item{Q30Day}{rolling means of the 30-day flow values of the designated date and previous 29 days (align="right")}
}
\description{
Adds selected n-day rolling means to a streamflow dataset. Based on selected n-days and alignment, the rolling mean for
  a given day is obtained by averaging the adjacent dates of daily mean values. For example, rolling days of '7' and 'right' alignment 
  would obtain a mean of the given and previous 6 days of daily mean flow. Rolling mean values will not be calculated if there is less 
  than the n-days provided.
}
\examples{
\dontrun{

add_rolling_means(data = flow_data, days = 7, align = 'centre')

add_rolling_means(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
