% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.stats.R
\name{annual_stats}
\alias{annual_stats}
\title{Compute multiple annual statistics.}
\usage{
annual_stats(station_name = NULL, flowdata = NULL, HYDAT = NULL,
  water_year = FALSE, start_year = NULL, end_year = NULL,
  exclude_years = NULL, basin_area = NA, zyp_trending = NA,
  zyp_alpha = 0.05, write_table = FALSE, write_transposed_table = FALSE,
  write_summary_table = FALSE, write_lowflow_table = FALSE,
  write_zyp_table = FALSE, write_zyp_plots = NA, report_dir = ".",
  na.rm = list(na.rm.global = FALSE), table_nddigits = 3)
}
\arguments{
\item{station_name}{Character. Identifier name of the stream or station. Required when supplying data through \code{flowdata}.
The station name will be used in plots and filenames of exported tables and plot. If using \code{HYDAT} to supply
data and no \code{station_name} is provided, the HYDAT station number will be the identifier.}

\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Q' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{water_year}{Logical. Set to \code{TRUE} if data should be summarized by water year (Oct-Sep) instead of the
default calendar year (Jan-Dec) (\code{water_year=FALSE}). Water years are designated by the year which they end in
(e.g. water year 2000 start on 1 Oct 1999 and ends on 30 Sep 2000).}

\item{start_year}{Numeric. The first year of streamflow data to analyze. If unset, the default \code{start_year} is the first
year of the data provided.}

\item{end_year}{Numeric. The last year of streamflow data to analyze. If unset, the default \code{end_year} is the last
year of the data provided.}

\item{basin_area}{Numeric. The upstream drainage basin area (in sq. km) of the station. Used to calculate runoff yields (mm).
If no value provided, yield calculations will result in NA values.}

\item{write_table}{Logical. Should a file be created with the calendar year computed percentiles?
The file name will be  \code{file.path(report_dir,paste(station_name,'-annual-cy-summary-stat.csv'))}.}

\item{write_transposed_table}{Logical. Should a file be created with the transposed of the annual statistics
(both calendar and water year)?
The file name will be  \code{file.path(report_dir,paste(station_name,'-annual-summary-stat-trans.csv'))}.}

\item{write_summary_table}{Logical. Should a file be created with a flow summary over the years between the
start_year and end_year (inclusive). This summary includes number of days, number of missing values,
mean, median, minimum, maximum, and standard deviation of \code{flowdata$Q}.
The file name will be \code{file.path(report_dir, paste(station_name,"-period-record-summary.csv", sep=""))}.}

\item{write_lowflow_table}{Logical. Should a file be created with the minimum value of \code{flowdata$Q} and date the
minimum occured.
The file name will be \code{file.path(report_dir,paste(station_name,"-lowflow-summary.csv",sep=""))}}

\item{report_dir}{Character. Folder location of where to write tables and plots. Default is the working directory.}

\item{na.rm}{TBD}

\item{table_nddigits}{Numeric. Number of significant digits to round the results in the written tables. Default is 3.}

\item{excluded.years}{Numeric. List of years to exclude final results from. Ex. 1990 or c(1990,1995:2000).}
}
\value{
A list with the following elements:
  \item{Qsummary}{Data frame with flow summary.}
  \item{Qstat_annual}{Data frame with summary statistics as listed at \code{\link{SummaryStatistics}}.}
  \item{Qstat_annual_tpose}{Data frame with transposed summary statistics as listed at \code{\link{SummaryStatistics}}.}
  \item{dates_missing_flows}{Data framw with dates of missing \code{flowdata$Q} between
         \code{start_year} and \code{end_year}}
  \item{file_Qstat_table}{Object with file name of *.csv file with calendar year summary statistics.}
  \item{file_Qstat_tpose_table}{Object with file name of *.csv file with transposed summary statistics.}
  \item{file_Qsummary_table}{Object with file name of *.csv file with summary statistics.}
  \item{file_Qlowflow_table}{Object with file name of *.csv file with low flow summary statistics.}
  \item{na.rm}{Missing value flags.}
  \item{Version}{Version of this function.}
  \item{Date}{Date function was run.}
}
\description{
Computes annual statistics (summary statistics, trending) of streamflow data.
Streamflow data can be supplied through the \code{flowdata} parameter or extracted from a 
HYDAT database using the tidyhydat package and \code{HYDAT} parameter.
and (optionally) saves the results in *.csv and *.pdf files.
}
\examples{
\dontrun{
stat.annual <- annual_stats(HYDAT="08HB048",
                            basin_area    = 10.1)

stat.annual <- annual_stats(station_name  ='Mission Creek',
                               flowdata     = flow,
                               water_year    = TRUE,
                               start_year    = 1967,
                               end_year      = 2014)
}
}
