% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_annual_cumulative_stats.R
\name{calc_annual_cumulative_stats}
\alias{calc_annual_cumulative_stats}
\title{Calculate annual (and seasonal) cumulative flows}
\usage{
calc_annual_cumulative_stats(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number = NULL, use_yield = FALSE,
  basin_area = NA, water_year = FALSE, water_year_start = 10,
  start_year = 0, end_year = 9999, exclude_years = NULL,
  incl_seasons = FALSE, transpose = FALSE)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{use_yield}{Logical value indicating whether to use yield runoff, in mm, instead of volumetric. Default \code{FALSE}.}

\item{basin_area}{Upstream drainage basin area, in square kilometres, to apply to observations. Three options:

(1) Leave blank if \code{groups} is STATION_NUMBER with HYDAT station numbers to extract basin areas from HYDAT.

(2) A single numeric value to apply to all observations.

(3) List each basin area for each group/station in groups (can override HYDAT value) as such \code{c("08NM116" = 795, 
"08NM242" = 10)}. Group not listed will result in NA basin area.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the calendar year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{incl_seasons}{Logical value indication whether to include seasonal yields and total discharges. Default \code{TRUE}.}

\item{transpose}{Logical value indicating if the results rows and columns are to be switched. Default \code{FALSE}.}
}
\value{
A tibble data frame with the following columns, ending with '_TotalQ_m3' or '_Yield_mm' based on selection:
  \item{Year}{calendar or water year selected}
  \item{Annual_*}{annual total flow, in m3 or mm}
  Default seasonal columns:
  \item{MMM-MMM_*}{first of two season total flows, in m3 or mm}
  \item{MMM-MMM_*}{second of two season total flows, in m3 or mm}
  \item{MMM-MMM_*}{first of four season total flows, in m3 or mm}
  \item{MMM-MMM_*}{second of four season total flows, in m3 or mm}
  \item{MMM-MMM_*}{third of four season total flows, in m3 or mm}
  \item{MMM-MMM_*}{fourth of four season total flows, in m3 or mm}
  Transposing data creates a column of 'Statistics' and subsequent columns for each year selected.
}
\description{
Calculates annual and seasonal total flows, volumetric or runoff yield flows, from a streamflow dataset. Calculates 
   the statistics from all daily discharge values from all years, unless specified. For water year and seasonal data, the designated
   year is the year in which the year or season ends. Two-seasons and four-seasons per year are calculated, with each 6 and 3-month
   seasons starting with the first month of the year (Jan for calendar year, specified for water year). Each season is designated
   by the calendar or water year in which it occurs.
}
\examples{
\dontrun{

calc_annual_cumulative_stats(station_number = "08NM116", 
                             water_year = TRUE, 
                             water_year_start = 8,
                             incl_seasons = TRUE)

}
}
