% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_annual_flow_timing.R
\name{calc_annual_flow_timing}
\alias{calc_annual_flow_timing}
\title{Calculate annual flow timing}
\usage{
calc_annual_flow_timing(flow_data = NULL, flow_dates = Date,
  flow_values = Value, flow_stations = STATION_NUMBER, HYDAT = NULL,
  percent_total = c(25, 33.3, 50, 75), water_year = FALSE,
  water_year_start = 10, start_year = 0, end_year = 9999,
  exclude_years = NULL, transpose = FALSE)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_dates}{a column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Leave blank if using \code{HYDAT} 
argument. Default \code{Date}.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{flow_stations}{a column in flow_data that contains station identifiers (names or numbers) for each flow data set, required if 
calculating results for multiple stations. Removing 'STATION_NUMBER' column in flow_data or incorrectly identifying will calculate 
statistics on all flow values from all stations. If using \code{HYDAT} argument, setting \code{flow_stations} to anything besides 
\code{STATION_NUMBER} will have similar effects. Default \code{STATION_NUMBER}.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{percent_total}{a numeric vector of percents of total annual flows to determine dates. Default \code{c(25,33.3,50,75)}.}

\item{water_year}{a logical value indicating whether to use water years to group flow data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{a numeric value indicating the month of the start of the water year. Used if \code{water_year=TRUE}. 
Default \code{10}.}

\item{start_year}{a numeric value of the first year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{a numeric value of the last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{a numeric vector of years to exclude from analysis. Leave blank if all years are required.}

\item{transpose}{a logical value indicating if the results rows and columns are to be switched. Default \code{FALSE}.}
}
\value{
A data frame with the following columns:
  \item{Year}{calendar or water year selected}
  \item{DoY_'n'pct_TotalQ}{day of year for each n-percent of total volumetric discharge}
  \item{Date_'n'pct_TotalQ}{date (YYYY-MM-DD) for each n-percent of total volumetric discharge}
  Default columns:
  \item{DoY_25pct_TotalQ}{day of year of 25-percent of total volumetric discharge}
  \item{Date_25pct_TotalQ}{date (YYYY-MM-DD) of 25-percent of total volumetric discharge}
  \item{DoY_33.3pct_TotalQ}{day of year of 33.3-percent of total volumetric discharge}
  \item{Date_33.3pct_TotalQ}{date (YYYY-MM-DD) of 33.3-percent of total volumetric discharge}
  \item{DoY_50pct_TotalQ}{day of year of 50-percent of total volumetric discharge}
  \item{Date_50pct_TotalQ}{date (YYYY-MM-DD) of 50-percent of total volumetric discharge}
  \item{DoY_75pct_TotalQ}{day of year of 75-percent of total volumetric discharge}
  \item{Date_75pct_TotalQ}{date (YYYY-MM-DD) of 75-percent of total volumetric discharge}
  Transposing data creates a column of "Statistics" (just DoY, not Date values) and subsequent columns for each year selected.
}
\description{
Calculates annual the timing (day of year) and date of occurrence of portions of total annual flow of daily flow 
   values from a streamflow dataset. Calculates the statistics from all daily discharge values from all years, unless specified.
}
\examples{
\dontrun{

calc_annual_flow_timing(flow_data = flow_data)

calc_annual_flow_timing(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8, percent_total = 50)

}
}
