% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_annual_lowflows.R
\name{calc_annual_lowflows}
\alias{calc_annual_lowflows}
\title{Calculate annual lowflows}
\usage{
calc_annual_lowflows(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, rolling_days = c(1, 3, 7, 30),
  rolling_align = "right", water_year = FALSE, water_year_start = 10,
  start_year = 0, end_year = 9999, exclude_years = NULL, months = 1:12,
  transpose = FALSE, ignore_missing = FALSE)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{groups}{Column in the \code{data} data frame that contains unique identifiers for different data sets. 
Only required if using the data frame option of \code{data} and groups column is not named 'STATION_NUMBER'.
Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column or identify 
another non-existing column name to remove this grouping. Identify another column if desired. Default \code{STATION_NUMBER}.}

\item{rolling_days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{rolling_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{months}{Numeric vector of months to include in analysis (ex. \code{6:8} for Jun-Aug). Leave blank to summarize 
all months (default \code{1:12}).}

\item{transpose}{Logical value indicating if the results rows and columns are to be switched. Default \code{FALSE}.}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{TRUE}.}
}
\value{
A data frame with the following columns:
  \item{Year}{calendar or water year selected}
  \item{Min_'n'_Day}{annual minimum for each n-day rolling mean, direction of mean specified by rolling_align}
  \item{Min_'n'_Day_DoY}{day of year for each annual minimum of n-day rolling mean}
  \item{Min_'n'_Day_Date}{date (YYYY-MM-DD) for each annual minimum of n-day rolling mean}
  Default columns:
  \item{Min_1_Day}{annual 1-day mean minimum (rolling_align=right)}
  \item{Min_1_Day_DoY}{day of year of annual 1-day mean minimum}
  \item{Min_1_Day_Date}{date (YYYY-MM-DD) of annual 1-day mean minimum}
  \item{Min_3_Day}{annual 3-day mean minimum (rolling_align=right)}
  \item{Min_3_Day_DoY}{day of year of annual 3-day mean minimum}
  \item{Min_3_Day_Date}{date (YYYY-MM-DD) of annual 3-day mean minimum}   
  \item{Min_7_Day}{annual 7-day mean minimum (rolling_align=right)}
  \item{Min_7_Day_DoY}{day of year of annual 7-day mean minimum}
  \item{Min_7_Day_Date}{date (YYYY-MM-DD) of annual 7-day mean minimum}
  \item{Min_30_Day}{annual 30-day mean minimum (rolling_align=right)}
  \item{Min_30_Day_DoY}{day of year of annual 30-day mean minimum}
  \item{Min_30_Day_Date}{date (YYYY-MM-DD) of annual 30-day mean minimum}
  Transposing data creates a column of "Statistics" and subsequent columns for each year selected. "Date" statistics
  not transposed.
}
\description{
Calculates annual n-day minimum values, and the day of year and date of occurrence of daily flow values from a 
   streamflow dataset. Calculates the statistics from all daily discharge values from all years, unless specified.
}
\examples{
\dontrun{

calc_annual_lowflows(data = "08NM116", water_year = TRUE, water_year_start = 8, rolling_days = c(3,7))

}
}
