% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_annual_stats.R
\name{calc_annual_stats}
\alias{calc_annual_stats}
\title{Calculate annual summary statistics}
\usage{
calc_annual_stats(flow_data = NULL, flow_dates = Date,
  flow_values = Value, flow_stations = STATION_NUMBER, HYDAT = NULL,
  percentiles = c(10, 90), rolling_days = 1, rolling_align = "right",
  water_year = FALSE, water_year_start = 10, start_year = 0,
  end_year = 9999, exclude_years = NULL, months = 1:12,
  transpose = FALSE, ignore_missing = FALSE)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_dates}{a column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Leave blank if using \code{HYDAT} 
argument. Default \code{Date}.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{flow_stations}{a column in flow_data that contains station identifiers for each flow data set.
Removing 'STATION_NUMBER' column in flow_data or incorrectly identifying will calculate statistics on all flow values from all stations.
If using \code{HYDAT} argument, setting \code{flow_stations} to anything besides \code{STATION_NUMBER} will have similar effects.
Default \code{STATION_NUMBER}.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{percentiles}{a numeric vector of percentiles to calculate. Set to NA if none required. Default \code{c(10,90)}.}

\item{rolling_days}{a numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{rolling_align}{a character identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{a logical value indicating whether to use water years to group flow data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{a numeric value indicating the month of the start of the water year. Used if \code{water_year=TRUE}. 
Default \code{10}.}

\item{start_year}{a numeric value of the first year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{a numeric value of the last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{a numeric vector of years to exclude from analysis. Leave blank if all years are required.}

\item{months}{a numeric vector of months to include in analysis (ex. \code{6:8} for Jun-Aug). Leave blank to summarize 
all months (default \code{1:12}).}

\item{transpose}{a logical value indicating if the results rows and columns are to be switched. Default \code{FALSE}.}

\item{ignore_missing}{a logical value indicating whether dates with missing flow values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{TRUE}.}
}
\value{
A data frame with the following columns:
  \item{Year}{calendar or water year selected}
  \item{Mean}{annual mean of all daily flows for a given year}
  \item{Median}{annual median of all daily flows for a given year}
  \item{Maximum}{annual maximum of all daily flows for a given year}
  \item{Minimum}{annual minimum of all daily flows for a given year}
  \item{P'n'}{each annual n-th percentile selected of all daily flows}
  Default percentile columns:
  \item{P10}{annual 10th percentile of all daily flows for a given year}
  \item{P90}{annual 90th percentile of all daily flows for a given year}
  Transposing data creates a column of "Statistics" and subsequent columns for each year selected.
}
\description{
Calculates annual monthly mean, median, maximum, minimum, and percentiles of daily flow values from a streamflow 
   dataset. Calculates the statistics from all daily discharge values from all years, unless specified.
}
\examples{
\dontrun{

calc_annual_stats(flow_data = flow_data)

calc_annual_stats(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8, percentiles = c(1:10))

calc_annual_stats(HYDAT = "08NM116", months = 7:9)

}
}
