% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_daily_stats.R
\name{calc_daily_stats}
\alias{calc_daily_stats}
\title{Calculate the daily summary statistics}
\usage{
calc_daily_stats(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, percentiles = c(5, 25, 75, 95), roll_days = 1,
  roll_align = "right", water_year = FALSE, water_year_start = 10,
  start_year = 0, end_year = 9999, exclude_years = NULL,
  complete_years = FALSE, transpose = FALSE, ignore_missing = TRUE)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{groups}{Column in the \code{data} data frame that contains unique identifiers for different data sets. 
Only required if using the data frame option of \code{data} and groups column is not named 'STATION_NUMBER'.
Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column or identify 
another non-existing column name to remove this grouping. Identify another column if desired. Default \code{STATION_NUMBER}.}

\item{percentiles}{Numeric vector of percentiles to calculate. Set to NA if none required. Default \code{c(5,25,75,95)}.}

\item{roll_days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{complete_years}{Logical values indicating whether to include only years with complete data in analysis. Default \code{FALSE}.}

\item{transpose}{Logical value indicating if the results rows and columns are to be switched. Default \code{FALSE}.}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{TRUE}.}
}
\value{
A data frame with the following columns:
  \item{Date}{date (MMM-DD) of daily statistics}
  \item{DayofYear}{day of year of daily statistics}
  \item{Mean}{daily mean of all flows for a given day of the year}
  \item{Median}{daily mean of all flows for a given day of the year}
  \item{Maximum}{daily mean of all flows for a given day of the year}
  \item{Minimum}{daily mean of all flows for a given day of the year}
  \item{P'n'}{each daily n-th percentile selected of all flows for a given day of the year}
  Default percentile columns:
  \item{P5}{daily 5th percentile of all flows for a given day of the year}
  \item{P25}{daily 25th percentile of all flows for a given day of the year}
  \item{P75}{daily 75th percentile of all flows for a given day of the year}
  \item{P95}{daily 95th percentile of all flows for a given day of the year}
  Transposing data creates a column of "Statistics" and subsequent columns for each year selected.
}
\description{
Calculates the daily mean, median, maximum, minimum, and percentiles for each day of the year of daily flow values 
   from a streamflow dataset. Calculate the statistics from all daily discharge values from all years, unless specified. Can determine
   statistics of rolling mean days (e.g. 7-day flows) using the roll_days argument. Note that statistics are based on the numeric days
   of year (1-365) and not the date (Jan1 - Dec 31) of year so day of year values for days after Feb 29 in leap years will be one value
   higher than non-leap years.
}
\examples{
\dontrun{


calc_daily_stats(data = "08NM116", 
                 water_year = TRUE, 
                 water_year_start = 8, 
                 percentiles = c(1:10))

}
}
