% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_longterm_percentile.R
\name{calc_longterm_percentile}
\alias{calc_longterm_percentile}
\title{Calculate the long-term percentiles}
\usage{
calc_longterm_percentile(data, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number, percentiles, roll_days = 1,
  roll_align = "right", water_year_start = 1, start_year, end_year,
  exclude_years, complete_years = FALSE, months = 1:12,
  transpose = FALSE)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{percentiles}{Numeric vector of percentiles to calculate. Required.}

\item{roll_days}{Numeric value of the number of days to apply a rolling mean. Default \code{1}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first 
('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year for analysis. Default \code{1}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{complete_years}{Logical values indicating whether to include only years with complete data in analysis. Default \code{FALSE}.}

\item{months}{Numeric vector of months to include in analysis (e.g.. \code{6:8} for Jun-Aug). Leave blank to summarize 
all months (default \code{1:12}).}

\item{transpose}{Logical value indicating if the results rows and columns are to be switched. Default \code{FALSE}.}
}
\value{
A tibble data frame of a long-term percentile of selected years and months.
}
\description{
Calculates the long-term percentiles discharge of a streamflow dataset. Averages all daily discharge values from all years,
  unless specified.
}
\examples{
\dontrun{

# Calculate the 20th percentile
calc_longterm_percentile(station_number = "08NM116",
                         percentile = 20)
                         
# Calculate the 90th percentile value with custom years
calc_longterm_percentile(station_number = "08NM116", 
                         start_year = 1980, 
                         end_year = 2010, 
                         percentile = 90)
}
}
