% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lt_percentile.R
\name{calc_lt_percentile}
\alias{calc_lt_percentile}
\title{Calculate the long-term percentiles}
\usage{
calc_lt_percentile(flowdata = NULL, HYDAT = NULL, percentiles = NA,
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, months = 1:12)
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{percentiles}{Numeric. Long-term percentile(s) to calculate. Required.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{months}{Integer. Vector of months to combine to summarize (ex. \code{6:8} for Jun-Aug). Leave blank for all months. 
Default \code{1:12}.}
}
\value{
A numeric value of a long-term percentile of selected years and months.
}
\description{
Calculates the long-term percentiles discharge of a streamflow dataset. Averages all daily discharge values from all years,
  unless specified.
}
\examples{
\dontrun{

calc_lt_percentile(flowdata = data, start_year = 1980, end_year = 2010, percentile = 90)

calc_lt_percentile(HYDAT = "08NM116", percentile = 20)

}
}
