% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_full_analysis.R
\name{compute_full_analysis}
\alias{compute_full_analysis}
\title{Computes a suite of tables and plots from various fasstr functions}
\usage{
compute_full_analysis(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number = NULL, sections = 1:7,
  basin_area = NA, water_year_start = 1, start_year = 0,
  end_year = 3000, exclude_years = NULL, ignore_missing = FALSE,
  zyp_method = "yuepilon", zyp_alpha = NA, write_to_dir = FALSE,
  foldername = NULL, plot_filetype = "pdf")
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{sections}{Sections of analysis to run (default is all (\code{1:7})):
\itemize{
 \item{1: Screening}
 \item{2: Long-term}
 \item{3: Annual}
 \item{4: Monthly}
 \item{5: Daily}
 \item{6: Annual Trends}
 \item{7: Low-flow Frequencies}
 }}

\item{basin_area}{Upstream drainage basin area, in square kilometres, to apply to observations. Three options:

(1) Leave blank if \code{groups} is STATION_NUMBER with HYDAT station numbers to extract basin areas from HYDAT.

(2) A single numeric value to apply to all observations.

(3) List each basin area for each group/station in groups (can override HYDAT value) as such \code{c("08NM116" = 795, 
"08NM242" = 10)}. Group not listed will result in NA basin area.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year for analysis. Default \code{1}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{FALSE}.}

\item{zyp_method}{Character string identifying the prewhitened trend method to use from 'zyp', either "zhang' or "yuepilon". 
Only required if section 7 is included. Default \code{"yuepilon"}.}

\item{zyp_alpha}{Numeric value of the significance level (ex. 0.05) of when to plot a trend line. Leave blank for no line.}

\item{write_to_dir}{Logical value indicating if all results are to also be written into a directory. Default \code{FALSE}.}

\item{foldername}{Name of folder to create on disk (if it does not exist) to create all folders and save tables and plots.}

\item{plot_filetype}{Image type to write. One of "png", "eps", "ps", "tex", "pdf", "jpeg", "tiff", "bmp", or "svg".
If type is not "pdf" then individual plots will be created instead of a combined PDF. Default \code{"pdf"}.}
}
\value{
A list of lists of tibble data frames and ggplot2 objects from various fasstr functions
   organized by the sections as listed above.
}
\description{
Calculates tables and plots of a suite of statistics from various fasstr functions. The statistics are grouped into
   7 sections (see below) which are stored in lists in the created object. Due to the number of tables and plots to be made, this 
   function may take several minutes to complete. Can also directly write all plots and tables directly into a directory
   ('foldername' argument) using the 'write_to_dir' argument.
   If using \code{ignore_missing = FALSE} (default) and there is missing data, some tables and plots may be empty
   and produce warnings. Use \code{ignore_missing = TRUE} to ignore the missing values or filter your data to complete years.
}
\examples{
\dontrun{

compute_full_analysis(station_number = "08NM116")

}
}
