% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr-package.R
\docType{package}
\name{fasstr-package}
\alias{fasstr}
\alias{fasstr-package}
\title{fasstr: Tools for Streamflow Data Analyses, Summaries, and Visualization}
\description{
Provides functions to summarize, analyze, and visualize streamflow data.
}
\references{
Tidyhydat information:


To download the latest version of hydat please:
 \itemize{
  \item use the \code{download_hydat()} function.
  }


 For more information on HYDAT 
 \itemize{
   \item Please see this description of the database: goo.gl/H3NXJQ
   \item This page is landing page for technical description of HYDAT:
   \url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/}
   \item This page links to a document that outlines database table definitions:
   \url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/HYDAT_Definition_EN.pdf}
   }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/fasstr}
  \item Report bugs at \url{https://github.com/bcgov/fasstr/issues}
}

}
\author{
\strong{Maintainer}: Jon Goetz \email{jon.goetz@gov.bc.ca}

Authors:
\itemize{
  \item Carl James Schwarz \email{cschwarz.stat.sfu.ca@gmail.com}
}

Other contributors:
\itemize{
  \item Robin Pike \email{robin.g.pike@gov.bc.ca} [contributor]
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
