% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_add_cumulative_volume.R
\name{fasstr_add_cumulative_volume}
\alias{fasstr_add_cumulative_volume}
\title{Add daily cumulative volumetric flows on an annual basis}
\usage{
fasstr_add_cumulative_volume(flowdata = NULL, HYDAT = NULL,
  water_year = FALSE, water_year_start = 10)
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}
}
\value{
A column of cumulative volumetric flows for each day for each year specific called 'Cumul_Volume' added to the flowdata 
   data frame input or HYDAT dataset, in units of cubic metres.
}
\description{
Add a column of daily cumulative volumetric flows on an annual basis to a streamflow dataset. Adds the volumetric discharge 
   from each day with the previous day(s) for each year, in units of cubic metres. The cumulative flows restart every year.
}
\examples{
\dontrun{

fasstr_add_cumulative_volume(flowdata = flowdata)

fasstr_add_cumulative_volume(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
