% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_add_cumulative_yield.R
\name{fasstr_add_cumulative_yield}
\alias{fasstr_add_cumulative_yield}
\title{Add cumulative volumetric flows on an annual basis.}
\usage{
fasstr_add_cumulative_yield(flowdata = NULL, HYDAT = NULL,
  water_year = FALSE, water_year_start = 10, basin_area = NA)
}
\arguments{
\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Value' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{water_year}{Logical (TRUE/FALSE). Choose to fill to the start of the first/last water years.}

\item{water_year_start}{Numeric. Month to start water year (1 to 12 for Jan to Dec). Default 10 (Oct).}

\item{basin_area}{Numeric. The upstream drainage basin area (in sq. km) of the station. Used to calculate runoff yields (mm)}
}
\description{
Add cumulative volumetric flows on an annual basis.
}
\examples{
\dontrun{

set example :)
}
}
