% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_add_daily_yield.R
\name{fasstr_add_daily_yield}
\alias{fasstr_add_daily_yield}
\title{Add volume per day}
\usage{
fasstr_add_daily_yield(flowdata = NULL, HYDAT = NULL, basin_area = NA)
}
\arguments{
\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Value' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{basin_area}{Numeric. The upstream drainage basin area (in sq. km) of the station. Used to calculate runoff yields (mm)}
}
\description{
Add cumulative volumetric flows on an annual basis.
}
\examples{
\dontrun{

set example :)
}
}
