% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_add_rolling_means.R
\name{fasstr_add_rolling_means}
\alias{fasstr_add_rolling_means}
\title{Add rolling means.}
\usage{
fasstr_add_rolling_means(flowdata = NULL, HYDAT = NULL, days = c(3, 7,
  30), align = "right")
}
\arguments{
\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Value' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{days}{Numeric. The number of days to apply the rolling mean. Default c(3,7,30).}

\item{align}{Character. specifyies whether the index of the result should be left- or right-aligned or centered 
(default) compared to the rolling window of observations}
}
\description{
Adds n-day rolling means. Ensures rolling means
}
\examples{
\dontrun{

set example :)
}
}
