% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_annual_days_outside_normal.R
\name{fasstr_annual_days_outside_normal}
\alias{fasstr_annual_days_outside_normal}
\title{Calculate annual days above and below normal}
\usage{
fasstr_annual_days_outside_normal(flowdata = NULL, HYDAT = NULL,
  lower_percentile = 25, upper_percentile = 75, water_year = FALSE,
  water_year_start = 10, start_year = NULL, end_year = NULL,
  exclude_years = NULL, transpose = FALSE, station_name = NA,
  write_table = FALSE, write_dir = ".")
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{transpose}{Logical. Switch the rows and columns of the results table. Default \code{FALSE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_table}{Logical. Write the table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{lower_percentile=25}{Numeric. Percentile indicating the lower limit of the normal range. Default \code{25}.}

\item{upper_percentile=25}{Numeric. Percentile indicating the upper limit of the normal range. Default \code{75}.}
}
\value{
A data frame of annual days above and below normal
}
\description{
Calculates the number of days per year outside of the 'normal' range (typically between 25 and 75th percentiles) for
   each day of the year. Upper and lower-range percentiles are calcuated for each day of the year of from all years, and then each 
   daily flow value for each year is compared. All days above or below the normal range are included. Calculates the statistics 
   from all daily discharge values from all years, unless specified.
}
\examples{
\dontrun{

fasstr_annual_days_outside_normal(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

fasstr_annual_days_outside_normal(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
