% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_annual_freq_analysis.R
\name{fasstr_annual_freq_analysis}
\alias{fasstr_annual_freq_analysis}
\title{Perform a volume frequency analysis on annual statistics}
\usage{
fasstr_annual_freq_analysis(flowdata = NULL, HYDAT = NULL,
  HYDAT_peaks = NA, rolling_days = c(1, 3, 7, 30),
  rolling_align = "right", water_year = FALSE, water_year_start = 10,
  start_year = NULL, end_year = NULL, exclude_years = NULL,
  months = 1:12, use_log = FALSE, use_max = FALSE,
  prob_plot_position = c("weibull", "median", "hazen"),
  prob_scale_points = c(0.9999, 0.999, 0.99, 0.9, 0.5, 0.2, 0.1, 0.02, 0.01,
  0.001, 1e-04), fit_distr = c("PIII", "weibull"),
  fit_distr_method = ifelse(fit_distr == "PIII", "MOM", "MLE"),
  fit_quantiles = c(0.975, 0.99, 0.98, 0.95, 0.9, 0.8, 0.5, 0.2, 0.1, 0.05,
  0.01), station_name = NA, write_table_overview = FALSE,
  write_table_stats = FALSE, write_table_plotdata = FALSE,
  write_table_quantiles = FALSE, write_plot_frequency = FALSE,
  write_imgtype = c("pdf", "png"), write_imgsize = c(4, 6),
  write_digits = 3, write_dir = ".", na.rm = list(na.rm.global = TRUE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{HYDAT_peaks}{Character. "MAX" or "MIN" instantaneous peaks pulled from HYDAT for analysis. Leave blank if not required.}

\item{rolling_days}{Numeric. The number of days to apply a rolling mean. Default \code{1}.}

\item{rolling_align}{Character. Specifies whether the dates of the rolling mean should be specified by the first ('left'), last ('right),
or middle ('center') of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{months}{Integer. Vector of months to consider for analysis (ex. \code{6:8} for Jun-Aug). Leave blank if all months
are required. Default \code{1:12}.}

\item{use_log}{Transfrom to log-scale before analysis?}

\item{use_max}{Analyze the maximums rather than the minimums.}

\item{prob_plot_position}{Which plotting positions should be used in the frequency plots. Points are plotted
against  (i-a)/(n+1-a-b) where \code{i} is the rank of the value; \code{n} is the sample size and
\code{a} and \code{b} are defined as:
(a=0, b=0) for Weibull plotting positions;
(a=.2; b=.3) for Median plotting postions;
(a=.5; b=.5) for Hazen plotting positions.}

\item{prob_scale_points}{What points should be plotted along the $X$ axis in the frequency plot.}

\item{fit_distr}{Which distribution should be fit? PIII = Pearson Log III distribution; weibull=Weibull distribution.}

\item{fit_distr_method}{Which method used to fit the distribution. MOM=Method of moments; MLE=maximum likelihood estimation.}

\item{fit_quantiles}{Which quantiles should be estimated from the fitted distribution?}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_table_overview}{Logical. Write the overview table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_table_stats}{Logical. Write the annual statistics (used for analysis) table as a .csv file to specified directory. 
Default \code{FALSE}.}

\item{write_table_plotdata}{Logical. Write the plotting data (used for analysis) table as a .csv file to specified directory. 
Default \code{FALSE}.}

\item{write_table_quantiles}{Logical. Write the results table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_plot_frequency}{Logical. Write the frequency plot to specified directory. Default \code{FALSE}.}

\item{write_imgtype}{Character. One of "pdf","png","jpeg","tiff", or "bmp" image types to write the plot as. Default \code{"pdf"}.}

\item{write_imgsize}{Numeric. Height and width, respectively, of saved plot. Default \code{c(5,11)}.}

\item{write_digits}{Numeric. Number of significant digits to round the results in the written table. Default \code{3}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na.rm}{TBD}
}
\value{
A list with the following elements:
  \item{Q_stat}{Data frame with Computed annual summary statistics used in analysis}
  \item{plotdata}{Data frame with Co-ordinates used in frequency plot.}
  \item{freqplot}{ggplot2 object with frequency plot}
  \item{fit}{List of fitted objects from fitdistrplus.}
  \item{fitted_quantiles}{Data frame with fitted quantiles.}
  \item{overview}{Data frame overview of parameters used in analysis.}
}
\description{
Performs a volume frequency analysis on annual statistics from a streamflow dataset. Calculates the statistics from all 
daily discharge values from all years, unless specified. Results are similar to HEC-SSP. Defaults....TO BE FINISHED
}
\examples{
\dontrun{

fasstr_annual_freq_analysis(HYDAT="08NM116",
                            start_year = 1980,
                            end_year = 2010)
                            
}
}
