% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_annual_lowflows_plots.R
\name{fasstr_annual_lowflows_plots}
\alias{fasstr_annual_lowflows_plots}
\title{Plot the annual low flow minimums and the day they occur.}
\usage{
fasstr_annual_lowflows_plots(flowdata = NULL, HYDAT = NULL,
  station_name = "fasstr", water_year = FALSE, water_year_start = 10,
  start_year = NULL, end_year = NULL, exclude_years = NULL,
  rolling_days = c(1, 3, 7, 30), rolling_align = "right",
  write_plot = FALSE, plot_type = "pdf", report_dir = ".",
  na.rm = list(na.rm.global = FALSE))
}
\arguments{
\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Value' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{station_name}{Character. Identifier name of the stream or station. Required when supplying data through \code{flowdata}.
The station name will be used in plots and filenames of exported tables and plot. If using \code{HYDAT} to supply
data and no \code{station_name} is provided, the HYDAT station number will be the identifier.}

\item{water_year}{Logical. Set to \code{TRUE} if data should be summarized by water year (Oct-Sep) instead of the
default calendar year (Jan-Dec) (\code{water_year=FALSE}). Water years are designated by the year which they end in
(e.g. water year 2000 start on 1 Oct 1999 and ends on 30 Sep 2000).}

\item{water_year_start}{Numeric. Month to start water year (1 to 12 for Jan to Dec).}

\item{start_year}{Numeric. The first year of streamflow data to analyze. If unset, the default \code{start_year} is the first
year of the data provided.}

\item{end_year}{Numeric. The last year of streamflow data to analyze. If unset, the default \code{end_year} is the last
year of the data provided.}

\item{exclude_years}{Numeric. List of years to exclude final results from. Ex. 1990 or c(1990,1995:2000).}

\item{rolling_days}{Numeric. The rolling n-day mean. Default c(1,3,7,30).}

\item{rolling_align}{Character. Specifies whether the date of the means should be left- (first day) or right (last day)-aligned 
or centered (middle day). Default right.}

\item{write_plot}{Logical. Should a file be created with the calendar year computed percentiles?
The file name will be  \code{file.path(report_dir,paste(station_name,'-annual-cy-summary-stat.csv'))}.}

\item{plot_type}{Character. pdf, png, bmp, jpeg, tiff. Default pdf.}

\item{report_dir}{Character. Folder location of where to write tables and plots. Default is the working directory.}

\item{na.rm}{TBD}
}
\description{
Plot the annual low flow minimums and the day they occur.
Streamflow data can be supplied through the \code{flowdata} parameter or extracted from a 
HYDAT database using the tidyhydat package and \code{HYDAT} parameter.
No transpose due to dates.
}
\examples{
\dontrun{

Coming soon :)


}
}
