% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_annual_total_flows.R
\name{fasstr_annual_total_flows}
\alias{fasstr_annual_total_flows}
\title{Calculate annual summary statistics}
\usage{
fasstr_annual_total_flows(flowdata = NULL, HYDAT = NULL, basin_area = NA,
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, incl_seasons = TRUE,
  transpose = FALSE, station_name = NA, write_table = FALSE,
  write_digits = 3, write_dir = ".", na.rm = list(na.rm.global = FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{basin_area}{Numeric. The upstream drainage basin area (in sq. km) of the station. Used to calculate runoff yields (mm)
or middle ('center') of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{incl_seasons}{Logical. Include seasonal yields and total discharges. Default \code{TRUE}.}

\item{transpose}{Logical. Switch the rows and columns of the results table. Default \code{FALSE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_table}{Logical. Write the table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_digits}{Numeric. Number of significant digits to round the results in the written table. Default \code{3}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na.rm}{TBD}
}
\value{
A data frame of annual and seasonal total flows (volumetric and yield)
}
\description{
Calculates annual monthly mean, median, maximum, minimum, and percentiles of daily flow values from a streamflow 
   dataset. Calculates the statistics from all daily discharge values from all years, unless specified.
}
\examples{
\dontrun{

fasstr_annual_total_flows(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

fasstr_annual_total_flows(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
