% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_annual_trends_analysis.R
\name{fasstr_annual_trends_analysis}
\alias{fasstr_annual_trends_analysis}
\title{Calculate prewhitened nonlinear trends on streamflow data}
\usage{
fasstr_annual_trends_analysis(flowdata = NULL, HYDAT = NULL,
  trendsdata = NULL, zyp_method = NA, basin_area = NA,
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, annual_percentiles = c(10, 90),
  monthly_percentiles = c(10, 20), lowflow_days = c(1, 3, 7, 30),
  lowflow_align = "right", totalflow_seasons = TRUE,
  timing_percent = c(25, 33, 50, 75), normal_percentiles = c(25, 75),
  transpose = FALSE, station_name = NA, write_table_data = FALSE,
  write_table_results = FALSE, write_dir = ".", na.rm = list(na.rm.global
  = FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{trendsdata}{Data frame. Annual data with column names of years and rows of annual statistics. Leave blank if using \code{HYDAT}}

\item{zyp_method}{Character. The prewhitened trend method to use from 'zyp', either "zhang' or "yuepilon". Required.}

\item{basin_area}{Numeric. Upstream drainage basin area of the hydrometric station, in sq. km. Leave blank if \code{HYDAT} is used or 
a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the basin area will be extracted from HYDAT. 
Setting the basin area will replace the HYDAT basin area.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{annual_percentiles}{Numeric. Vector of percentiles to calculate annually. Set to NA if none required. Used for fasstr_annual_stats()
function. Default \code{c(10,90)}.}

\item{monthly_percentiles}{Numeric. Vector of percentiles to calculate monthly for each year. Set to NA if none required. Used for 
fasstr_monthly_stats() function. Default \code{c(10,90)}.}

\item{lowflow_days}{Numeric. The number of days to apply a rolling mean. Used for fasstr_annual_lowflows() function. Default \code{1}.}

\item{lowflow_align}{Character. Specifies whether the dates of the rolling mean should be specified by the first ('left'), last ('right),
or middle ('center') of the rolling n-day group of observations. Used for fasstr_annual_lowflows() function. Default \code{'right'}.}

\item{totalflow_seasons}{Logical. Include seasonal yields and total discharges.Used for fasstr_annual_total_flows() function. 
Default \code{TRUE}.}

\item{timing_percent}{Numeric. Percents of annual total flows to determine dates. Used for fasstr_annual_flow_timing() function. 
Default \code{c(25,33.3,50,75)}.}

\item{normal_percentiles}{Numeric. Lower and upper percentiles, respectively indicating the limits of the normal range. 
Used for fasstr_annual_days_outside_normal() function. Default \code{c(25,75)}.}

\item{transpose}{Logical. Switch the rows and columns of the results table. Default \code{FALSE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_table_data}{Logical. Write the trending input data table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_table_results}{Logical. Write the trending results table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na.rm}{TBD}
}
\value{
A data frame of results of a trends analysis
}
\description{
Calculates prewhitened nonlinear trends on annual streamflow data. Uses the
   \href{https://cran.r-project.org/web/packages/zyp/index.html}{'zyp'} package to trend. Review 'zyp' to understand its methology.
   Use \code{trendsdata} if providing your own data frame of statistics to trend. If using \code{HYDAT} or \code{flowdata}, then all
   annual statistics will be calculated using the fasstr_annual_all_stats() function which uses the following fasstr functions:
\itemize{
 \item{fasstr_annual_stats()}
 \item{fasstr_annual_lowflows()}
 \item{fasstr_annual_total_flows()}
 \item{fasstr_annual_flow_timing()}
 \item{fasstr_monthly_stats()}
 \item{fasstr_annual_days_outside_normal()}
 }
}
\examples{
\dontrun{

fasstr_annual_trends_analysis(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

fasstr_annual_trends_analysis(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
