% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_daily_stats_plots.R
\name{fasstr_daily_stats_plots}
\alias{fasstr_daily_stats_plots}
\title{Plot the daily summary statistics}
\usage{
fasstr_daily_stats_plots(flowdata = NULL, HYDAT = NULL, rolling_days = 1,
  rolling_align = "right", water_year = FALSE, water_year_start = 10,
  start_year = NULL, end_year = NULL, exclude_years = NULL,
  log_discharge = TRUE, station_name = NA, write_plot = FALSE,
  write_imgtype = "pdf", write_imgsize = c(4, 8.5), write_dir = ".",
  na.rm = list(na.rm.global = FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{rolling_days}{Numeric. The number of days to apply a rolling mean. Default \code{1}.}

\item{rolling_align}{Character. Specifies whether the dates of the rolling mean should be specified by the first ('left'), last ('right),
or middle ('center') of the rolling n-day group of observations. Default \code{'right'}.
by the year in which they end. Default \code{FALSE}.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{log_discharge}{Logical. Place the discharge axis (Y) on log scale. Default \code{TRUE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_plot}{Logical. Write the plot to specified directory. Default \code{FALSE}.}

\item{write_imgtype}{Character. One of "pdf","png","jpeg","tiff", or "bmp" image types to write the plot as. Default \code{"pdf"}.}

\item{write_imgsize}{Numeric. Height and width, respectively, of saved plot. Default \code{c(5,11)}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na.rm}{TBD}
}
\value{
A list of ggplot2 objects, the first the daily statistics plot containing the listed plots below, and the sebsequent plots for each
   year of data provided containing the first plot plus the daily flow data for each year.
  \item{Mean}{daily mean}
  \item{Median}{daily median}
  \item{25-75 Percentiles Range}{a ribbon showing the range of data between the daily 25th and 75th percentiles}
  \item{5-95 Percentiles Range}{a ribbon showing the range of data between the daily 5th and 95th percentiles}
  \item{Max-Min Range}{a ribbon showing the range of data between the daily minimum and maximums}
  \item{'Year' Flows}{(on annual plots) the daily flows for the designated year}
}
\description{
Plot the daily mean, median, maximum, minimum, and percentiles for each day of the year of daily flow values 
   from a streamflow dataset. Plots the statistics from all daily discharge values from all years, unless specified. Can determine
   statistics of rolling mean days (e.g. 7-day flows) using the rolling_days argument. Data calculated using fasstr_longterm_stats()
   function.
}
\examples{
\dontrun{

fasstr_daily_stats_plot(flowdata = flowdata, station_name = "MissionCreek", write_plot = TRUE)

fasstr_daily_stats_plot(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
