% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_flow_duration_curves.R
\name{fasstr_flow_duration_curves}
\alias{fasstr_flow_duration_curves}
\title{Plot flow duration curves}
\usage{
fasstr_flow_duration_curves(flowdata = NULL, HYDAT = NULL,
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, incl_months = 1:12,
  incl_longterm = TRUE, custom_months = NULL, log_discharge = TRUE,
  station_name = NA, write_plot = FALSE, write_imgtype = "pdf",
  write_imgsize = c(6, 11), write_dir = ".", na.rm = list(na.rm.global =
  FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{incl_months}{Numeric. Months of data to plot. NA if no months required. Default \code{1:12}.}

\item{incl_longterm}{Logical. Include longterm curve of all data. Default \code{TRUE}.}

\item{custom_months}{Integer. Vector of months to combine to plot (ex. \code{6:8} for Jun-Aug). Leave blank for no custom months to plot.}

\item{log_discharge}{Logical. Plot the discharge axis (Y-axis) on a logarithmic scale. Default \code{TRUE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_plot}{Logical. Write the plot to specified directory. Default \code{FALSE}.}

\item{write_imgtype}{Character. One of "pdf","png","jpeg","tiff", or "bmp" image types to write the plot as. Default \code{"pdf"}.}

\item{write_imgsize}{Numeric. Height and width, respectively, of saved plot. Default \code{c(6,11)}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na.rm}{TBD}
}
\value{
A plot of flow duration curves for specified long-term months and long-term flow data.
}
\description{
Plots flow duration curves, percent time a flow value is equalled or exceeded, for a streamflow dataset. Plots 
   statistics from all daily discharge values from all years, unless specified. Data calculated using fasstr_longterm_stats() 
   function then converted for plotting.
}
\examples{
\dontrun{

fasstr_flow_duration_curves(flowdata = flowdata, station_name = "MissionCreek", write_plot = TRUE)

fasstr_flow_duration_curves(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
