% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_longterm_stats.R
\name{fasstr_longterm_stats}
\alias{fasstr_longterm_stats}
\title{Calculate the long-term and long-term monthly summary statistics}
\usage{
fasstr_longterm_stats(flowdata = NULL, HYDAT = NULL, percentiles = c(10,
  90), water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, station_name = NA,
  transpose = FALSE, write_table = FALSE, write_digits = 3,
  write_dir = ".", na.rm = list(na.rm.global = FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{percentiles}{}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{transpose}{Logical. Switch the rows and columns of the results.}

\item{write_table}{Logical. Should a file be created with the calendar year computed percentiles?
The file name will be  \code{file.path(write_dir,paste(station_name,'-annual-cy-summary-stat.csv'))}.}

\item{write_digits}{Numeric. Number of significant digits to round the results in the written tables. Default is 3.}

\item{write_dir}{Character. Folder location of where to write tables and plots. Default is the working directory.}

\item{na.rm}{TBD}

\item{percentiles}{}
}
\description{
Calculates the long-term and long-term monthly mean, median, maximum, minimum, and percentiles of daily flow values 
   from a streamflow dataset. Calculate the statistics from all daily discharge values from all years, unless specified.
}
\examples{
\dontrun{

coming soon :)
}
}
