% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_write_daily_flows.R
\name{fasstr_write_daily_flows}
\alias{fasstr_write_daily_flows}
\title{Add calendar and water year date variables.}
\usage{
fasstr_write_daily_flows(flowdata = NULL, HYDAT = NULL,
  station_name = "fasstr", report_dir = ".", fill_missing_dates = FALSE,
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, exclude_rm = FALSE, na = NA)
}
\arguments{
\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Value' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{fill_missing_dates}{Logical.}

\item{water_year}{Logical. Set to \code{TRUE} if data should be summarized by water year (Oct-Sep) instead of the
default calendar year (Jan-Dec) (\code{water_year=FALSE}). Water years are designated by the year which they end in
(e.g. water year 2000 start on 1 Oct 1999 and ends on 30 Sep 2000).}

\item{water_year_start}{Numeric. Month to start water year (1 to 12 for Jan to Dec).}
}
\description{
Adds mulitple date variables to a dataframe from a column of dates, including
  year, month (numeric and text), day of years, and water years and day of water years.
}
\examples{
\dontrun{

set example :)
}
}
