% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longterm.stats.R
\name{longterm.stats}
\alias{longterm.stats}
\title{Compute long-term monthly and annual (calendar or water year) summary statistics.}
\usage{
longterm.stats(station.name = NULL, flow.data = NULL, HYDAT = NULL,
  water.year = FALSE, start.year = NULL, end.year = NULL,
  exclude.years = NULL, write.table = FALSE,
  write.transposed.table = FALSE, report.dir = ".", csv.nddigits = 3,
  na.rm = list(na.rm.global = TRUE))
}
\arguments{
\item{station.name}{Character. Identifier name of the stream or station. Required when supplying data through \code{flow.data}.
The station name will be used in plots and filenames of exported tables and plot. If using \code{HYDAT} to supply
data and no \code{station.name} is provided, the HYDAT station number will be the identifier.}

\item{flow.data}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Q' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flow.data} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{water.year}{Logical. Set to \code{TRUE} if data should be summarized by water year (Oct-Sep) instead of the
default calendar year (Jan-Dec) (\code{water.year=FALSE}). Water years are designated by the year which they end in
(e.g. water year 2000 start on 1 Oct 1999 and ends on 30 Sep 2000).}

\item{start.year}{Numeric. The first year of streamflow data to analyze. If unset, the default \code{start.year} is the first
year of the data provided.}

\item{end.year}{Numeric. The last year of streamflow data to analyze. If unset, the default \code{end.year} is the last
year of the data provided.}

\item{write.table}{Should a file be created with the computed statistics?
The file name will be  \code{file.path(report.dir,paste(station.name,'-longterm-summary-stat.csv'))}.}

\item{write.transposed.table}{Should a file be created with the transposed of the statistics report?
The file name will be  \code{file.path(report.dir,paste(station.name,'-longterm-summary-stat-trans.csv'))}.}

\item{report.dir}{Character. Folder location of where to write tables and plots. Default is the working directory.}

\item{csv.nddigits}{Numeric. Number of significant digits to round the results in the written tables. Default is 3.}

\item{na.rm}{TBD}

\item{excluded.years}{Numeric. List of years to exclude final results from. Ex. 1990 or c(1990,1995:2000).}
}
\value{
A list with the following elements:
  \item{Q.stat.longterm}{Data frame with the long-term statistics of \code{flow.data$Q} by month
        and overall between \code{start.year} and \code{end.year}}
  \item{Q.stat.longterm.trans}{Data frame with the long-term statistics of \code{flow.data$Q} transposed.}
  \item{file.stat.csv}{Object with file name of *.csv file with long term summary statistics.}
   \item{file.stat.trans.csv}{Object with file name of *.csv file with transposed long-term summary statistics.}
   \item{na.rm}{Missing value flags.}
   \item{Version}{Version of this function.}
   \item{Date}{Date function was run.}
}
\description{
Computes long-term summary statistics on \code{flow$Q} variable
   between \code{start.year} and \code{end.year} inclusive for calendar or water years.
   It (optionally) saves the results in *.csv and *.pdf files.
}
\examples{
\dontrun{
stat.longterm <- longterm.stats(
                         station.name  ='ABCDE',
                         flow.data          =flow,
                         start.year    =1960,
                         end.year      =2015)
}

}
