% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_flow_timing.R
\name{plot_annual_flow_timing}
\alias{plot_annual_flow_timing}
\title{Plot annual flow timing}
\usage{
plot_annual_flow_timing(flow_data = NULL, flow_dates = Date,
  flow_values = Value, HYDAT = NULL, percent_total = c(25, 33.3, 50, 75),
  water_year = FALSE, water_year_start = 10, start_year = 0,
  end_year = 9999, exclude_years = NULL)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_dates}{a column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Leave blank if using \code{HYDAT} 
argument. Default \code{Date}.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{percent_total}{a numeric vector of percents of total annual flows to determine dates. Default \code{c(25,33.3,50,75)}.}

\item{water_year}{a logical value indicating whether to use water years to group flow data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{a numeric value indicating the month of the start of the water year. Used if \code{water_year=TRUE}. 
Default \code{10}.}

\item{start_year}{a numeric value of the first year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{a numeric value of the last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{a numeric vector of years to exclude from analysis. Leave blank if all years are required.}
}
\value{
A ggplot2 object with the following plots:
  \item{DoY_'n'pct_TotalQ}{day of year for each n-percent of total volumetric discharge}
  Default plots:   
  \item{DoY_25pct_TotalQ}{day of year of 25-percent of total volumetric discharge}
  \item{DoY_33.3pct_TotalQ}{day of year of 33.3-percent of total volumetric discharge}
  \item{DoY_50pct_TotalQ}{day of year of 50-percent of total volumetric discharge}
  \item{DoY_75pct_TotalQ}{day of year of 75-percent of total volumetric discharge}
}
\description{
Plots annual the timing (day of year) of occurrence of portions of total annual flow of daily flow 
   values from a streamflow dataset. Calculates the statistics from all daily discharge values from all years, unless specified.
   Data calculated using calc_annual_flow_timing() function.
}
\examples{
\dontrun{

plot_annual_flow_timing(flowdata = flowdata)

plot_annual_flow_timing(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8, percent_total = 50)

}
}
