% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_outside_normal.R
\name{plot_annual_outside_normal}
\alias{plot_annual_outside_normal}
\title{Plot annual days above and below normal}
\usage{
plot_annual_outside_normal(data, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number, normal_percentiles = c(25,
  75), roll_days = 1, roll_align = "right", water_year_start = 1,
  start_year, end_year, exclude_years, months = 1:12,
  include_title = FALSE)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{normal_percentiles}{Numeric vector of two values, lower and upper percentiles, respectively indicating the limits of the 
normal range. Default \code{c(25,75)}.}

\item{roll_days}{Numeric value of the number of days to apply a rolling mean. Default \code{1}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first 
('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year for analysis. Default \code{1}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{months}{Numeric vector of months to include in analysis (e.g.. \code{6:8} for Jun-Aug). Leave blank to summarize 
all months (default \code{1:12}).}

\item{include_title}{Logical value to indicate adding the group/station number to the plot, if provided. Default \code{FALSE}.}
}
\value{
A list of ggplot2 objects with the following for each station provided:
  \item{Annual_Days_Outside_Normal}{a plot that contains the number of days outside normal}
  Default plots on each object:  
  \item{Days_Below_Normal}{number of days per year below the daily normal (default 25th percentile)}
  \item{Days_Above_Normal}{number of days per year above the daily normal (default 75th percentile)}
  \item{Days_Outside_Normal}{number of days per year below and above the daily normal (default 25/75th percentile)}
}
\description{
Plots the number of days per year outside of the 'normal' range (typically between 25 and 75th percentiles) for
   each day of the year. Upper and lower-range percentiles are calcuated for each day of the year of from all years, and then each 
   daily flow value for each year is compared. All days above or below the normal range are included. Calculates the statistics 
   from all daily discharge values from all years, unless specified. Data calculated using calc_annual_outside_normal() function.
}
\examples{
\dontrun{

# Plot statistics with default limits of normal (25 and 75th percentiles)
plot_annual_outside_normal(station_number = "08NM116") 

# Plot statistics with custom limits of normal
plot_annual_outside_normal(station_number = "08NM116",
                           normal_percentiles = c(10,90))
}
}
\seealso{
\code{\link{calc_annual_outside_normal}}
}
