% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_stats.R
\name{plot_annual_stats}
\alias{plot_annual_stats}
\title{Plot annual summary statistics}
\usage{
plot_annual_stats(flow_data = NULL, flow_dates = Date,
  flow_values = Value, HYDAT = NULL, percentiles = c(10, 90),
  rolling_days = 1, rolling_align = "right", water_year = FALSE,
  water_year_start = 10, start_year = 0, end_year = 9999,
  exclude_years = NULL, months = 1:12, log_discharge = FALSE,
  ignore_missing = TRUE)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_dates}{a column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Leave blank if using \code{HYDAT} 
argument. Default \code{Date}.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{percentiles}{a numeric vector of percentiles to plot. Set to NA if none required. Default \code{c(10,90)}.}

\item{water_year}{a logical value indicating whether to use water years to group flow data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{a numeric value indicating the month of the start of the water year. Used if \code{water_year=TRUE}. 
Default \code{10}.}

\item{start_year}{a numeric value of the first year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{a numeric value of the last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{a numeric vector of years to exclude from analysis. Leave blank if all years are required.}

\item{months}{a numeric vector of months to include in analysis (ex. \code{6:8} for Jun-Aug). Leave blank to summarize 
all months (default \code{1:12}).}

\item{log_discharge}{a logical value to indicate plotting thedischarge axis (Y-axis) on a logarithmic scale. Default \code{TRUE}.}

\item{ignore_missing}{a logical value indicating whether dates with missing flow values should be included in the calculation for
plotting. If \code{TRUE} then a statistic will be plotted regardless of missing dates. If \code{FALSE} then only statistics 
from time periods with no missing dates will be returned. Default \code{TRUE}.}
}
\value{
A ggplot2 object with the following plots (percentile plots optional):
  \item{Mean}{annual mean of all daily flows}
  \item{Median}{annual median of all daily flows}
  \item{Maximum}{annual maximum of all daily flows}
  \item{Minimum}{annual minimum of all daily flows}
}
\description{
Plots annual monthly mean, median, maximum, minimum, and percentiles of daily flow values from a streamflow 
   dataset. Plots the statistics from all daily discharge values from all years, unless specified. Data 
   calculated using calc_annual_stats() function.
}
\examples{
\dontrun{

plot_annual_stats(flowdata = flowdata)

plot_annual_stats(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

plot_annual_stats(HYDAT = "08NM116", months = 7:9)

}
}
