% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_total_flows.R
\name{plot_annual_total_flows}
\alias{plot_annual_total_flows}
\title{Plot annual and seasonal total flows}
\usage{
plot_annual_total_flows(flowdata = NULL, HYDAT = NULL, basin_area = NA,
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, station_name = NA,
  write_plot = FALSE, write_imgtype = "pdf", write_imgsize = c(5, 8.5),
  write_dir = ".", na.rm = list(na.rm.global = FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{basin_area}{Numeric. The upstream drainage basin area (in sq. km) of the station. Used to calculate runoff yields (mm)
or middle ('center') of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_plot}{Logical. Write the plot to specified directory. Default \code{FALSE}.}

\item{write_imgtype}{Character. One of "pdf","png","jpeg","tiff", or "bmp" image types to write the plot as. Default \code{"pdf"}.}

\item{write_imgsize}{Numeric. Height and width, respectively, of saved plot. Default \code{c(5,11)}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}
}
\value{
A list of the following ggplot2 objects (yield plots unavailable if no basin_area):
  \item{TotalQ_Annual}{ggplot2 object of annual total volumetric discharge, in cubic metres}
  \item{TotalQ_Two_Seasons}{ggplot2 object of Oct-Mar and Apr-Sep total volumetric discharges, in cubic metres}
  \item{TotalQ_Four_Seasons}{ggplot2 object of Jan-Mar, Apr-Jun, Jul-Sep, Oct-Dec total volumetric discharges, in cubic metres}
  \item{Yield_Annual}{ggplot2 object of annual runoff yield, in millimetres}
  \item{Yield_Two_Seasons}{ggplot2 object of Oct-Mar and Apr-Sep runoff yields, in millimetres}
  \item{Yield_Four_Seasons}{ggplot2 object of Jan-Mar, Apr-Jun, Jul-Sep, Oct-Dec runoff yields, in millimetres}
}
\description{
Plots annual and seasonal total flows, volumetric and runoff yield flows, from a streamflow dataset. Calculates 
   the statistics from all daily discharge values from all years, unless specified. Data calculated from
   calc_annual_total_flows() function. For water year and seasonal data, the designated
   year is the year in which the year or season ends. For example, if using water years with a start month of 11, the OND season is
   designated by the water year which starts in November (designated by the calendar year in which it ends).
}
\examples{
\dontrun{

plot_annual_total_flows(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

plot_annual_total_flows(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
