% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_trends.R
\name{plot_annual_trends}
\alias{plot_annual_trends}
\title{Plot prewhitened nonlinear trends on streamflow data}
\usage{
plot_annual_trends(trends_results = NULL, zyp_alpha = 0.05)
}
\arguments{
\item{trends_results}{Data frame of results from the compute_annual_trends() function, with original data included
(incl_data = TRUE). The first column must be 'Statistics' with the following columns with results and then data.}

\item{zyp_alpha}{Numeric value of the significance level of when to plot a trend line. Set to NA if no line required.}
}
\value{
A list of ggplot2 objects with plots of trended annual statistics, as the name of each object, with a trend line
   plotted if less than the zyp_alpha provided.
}
\description{
Plots the annual data and significant trend from the results from all statistics from a compute_annual_trends()
   analysis.
}
\examples{
\dontrun{

trends <- compute_annual_trends("08NM116",
                                zyp_method = "yuepilon",
                                start_year = 1980, end_year = 2010)
 
plot_annual_trends(trends_results = trends,
                   zyp_alpha = 0.05)

}
}
