% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily_cumulative_stats.R
\name{plot_daily_cumulative_stats}
\alias{plot_daily_cumulative_stats}
\title{Plot cumulative daily flow statistics}
\usage{
plot_daily_cumulative_stats(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, use_yield = FALSE, basin_area = NA,
  water_year = FALSE, water_year_start = 10, start_year = 0,
  end_year = 9999, exclude_years = NULL, complete_years = FALSE,
  ignore_missing = TRUE, log_discharge = FALSE, include_year = NULL)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{groups}{Column in the \code{data} data frame that contains unique identifiers for different data sets. 
Only required if using the data frame option of \code{data} and groups column is not named 'STATION_NUMBER'.
Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column or identify 
another non-existing column name to remove this grouping. Identify another column if desired. Default \code{STATION_NUMBER}.}

\item{use_yield}{Logical value indicating whether to use yield runoff, in mm, instead of volumetric. Default \code{FALSE}.}

\item{basin_area}{Upstream drainage basin area to apply to daily observations. Options:

Leave blank if \code{groups} is STATION_NUMBER with HYDAT station numbers to extract basin areas from HYDAT.

Single numeric value to apply to all observations.

List each basin area for each grouping factor (can override HYDAT value) as such \code{c("08NM116" = 795, "08NM242" = 10)}.
Factors not listed will result in NA basin areas.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{complete_years}{Logical values indicating whether to include only years with complete data in analysis. Default \code{FALSE}.}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{TRUE}.}

\item{log_discharge}{Logical. Place the discharge axis (Y) on log scale. Default \code{FALSE}.}
}
\value{
A list of ggplot2 objects, the first the daily cumulative statistics plot containing the listed plots below, and the sebsequent 
   plots for each year of data provided containing the first plot plus the daily cumulative flow data for each year. Default units in
   cubic metres, millimetres if use_yield and basin_area provided.
  \item{Mean}{daily cumulative mean}
  \item{Median}{daily cumulative median}
  \item{Min-5 Percentile Range}{a ribbon showing the range of data between the daily cumulative minimum and 5th percentile}
  \item{5-25 Percentiles Range}{a ribbon showing the range of data between the daily cumulative 5th and 25th percentiles}
  \item{25-75 Percentiles Range}{a ribbon showing the range of data between the daily cumulative 25th and 75th percentiles}
  \item{75-95 Percentiles Range}{a ribbon showing the range of data between the daily cumulative 75th and 95th percentiles}
  \item{95 Percentile-Max Range}{a ribbon showing the range of data between the daily cumulative 95th percentile and the maximum}
  \item{'Year' Flows}{(on annual plots) the daily cumulative flows for the designated year}
}
\examples{
\dontrun{

plot_daily_cumulative_stats(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
