% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily_stats.R
\name{plot_daily_stats}
\alias{plot_daily_stats}
\title{Plot the daily summary statistics}
\usage{
plot_daily_stats(data = NULL, dates = Date, values = Value,
  rolling_days = 1, rolling_align = "right", water_year = FALSE,
  water_year_start = 10, start_year = 0, end_year = 9999,
  exclude_years = NULL, complete_years = FALSE, ignore_missing = TRUE,
  log_discharge = TRUE, include_year = NULL)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{rolling_days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{rolling_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{complete_years}{Logical values indicating whether to include only years with complete data in analysis. Default \code{FALSE}.}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{TRUE}.}

\item{log_discharge}{Logical value to indicate plotting the discharge axis (Y-axis) on a logarithmic scale. Default \code{TRUE}.}

\item{include_year}{A numeric value indicating a year of daily flows to add to the daily statistics plot. Leave blank for no years.}

\item{percentiles}{Numeric vector of percentiles to calculate. Set to NA if none required. Default \code{NA}.}
}
\value{
A list of ggplot2 objects, the first the daily statistics plot containing the listed plots below, and the sebsequent plots for each
   year of data provided containing the first plot plus the daily flow data for each year.
  \item{Mean}{daily mean}
  \item{Median}{daily median}
  \item{25-75 Percentiles Range}{a ribbon showing the range of data between the daily 25th and 75th percentiles}
  \item{5-95 Percentiles Range}{a ribbon showing the range of data between the daily 5th and 95th percentiles}
  \item{Max-Min Range}{a ribbon showing the range of data between the daily minimum and maximums}
  \item{'Year' Flows}{(on annual plots) the daily flows for the designated year}
}
\description{
Plot the daily mean, median, maximum, minimum, and percentiles for each day of the year of daily flow values 
   from a streamflow dataset. Plots the statistics from all daily discharge values from all years, unless specified. Can determine
   statistics of rolling mean days (e.g. 7-day flows) using the rolling_days argument. Data calculated using calc_daily_stats()
   function.
}
\examples{
\dontrun{

plot_daily_stats(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
