% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_screening.R
\name{plot_data_screening}
\alias{plot_data_screening}
\title{Plot annual summary statistics for data screening}
\usage{
plot_data_screening(data = NULL, dates = Date, values = Value,
  rolling_days = 1, rolling_align = "right", water_year = FALSE,
  water_year_start = 10, start_year = 0, end_year = 9999)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{rolling_days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{rolling_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}
}
\value{
A ggplot2 object with the following plots:
  \item{Minimum}{annual minimum of all daily flows for a given year}
  \item{Maximum}{annual maximum of all daily flows for a given year}
  \item{Mean}{annual mean of all daily flows for a given year}
  \item{StandardDeviation}{annual 1 standard deviation of all daily flows for a given year}
}
\description{
Plots the mean, median, maximum, minimum, standard deviation of annual flows. Plots the statistics from all daily 
   discharge values from all years, unless specified. Data calculated using screen_flow_data() function.
}
\examples{
\dontrun{

plot_data_screening(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
