% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_longterm_stats.R
\name{plot_longterm_stats}
\alias{plot_longterm_stats}
\title{Plot the long-term and long-term monthly summary statistics}
\usage{
plot_longterm_stats(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number = NULL, roll_days = 1,
  roll_align = "right", water_year = FALSE, water_year_start = 10,
  start_year = 0, end_year = 9999, exclude_years = NULL,
  complete_years = FALSE, ignore_missing = FALSE, log_discharge = TRUE,
  include_title = FALSE)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{roll_days}{Numeric value of the number of days to apply a rolling mean. Default \code{1}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first 
('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the calendar year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{complete_years}{Logical values indicating whether to include only years with complete data in analysis. Default \code{FALSE}.}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{FALSE}.}

\item{log_discharge}{Logical value to indicate plotting the discharge axis (Y-axis) on a logarithmic scale. Default \code{FALSE}.}

\item{include_title}{Logical value to indicate adding the group/station number to the plot, if provided. Default \code{FALSE}.}
}
\value{
A list of ggplot2 objects with the following for each station provided:
  \item{Long-term_Stats}{a plot that contains long-term flow statistics}
  Default plots on each object:  
  \item{Long-term Mean}{mean of all daily flows over all years}
  \item{Long-term Median}{median of all daily flows over all years}
  \item{Monthly Mean}{mean of all daily flows for each month over all years}
  \item{Monthly Median}{median of all daily flows for each month over all years}
  \item{25-75 Percentiles Range}{a ribbon showing the range of data between the monthly 25th and 75th percentiles}
  \item{5-95 Percentiles Range}{a ribbon showing the range of data between the monthly 5th and 95th percentiles}
  \item{Max-Min Range}{a ribbon showing the range of data between the monthly minimum and maximums}
}
\description{
Plots the long-term and long-term monthly mean, median, maximum, minimum, and 5, 25, 75, and 95 percentiles of daily 
   flow values from a single streamflow dataset. Plots statistics from all daily discharge values from all years, unless specified. 
   Data calculated using calc_longterm_stats() function.
}
\examples{
\dontrun{

plot_longterm_stats(station_number = "08NM116", 
                    water_year = TRUE, 
                    water_year_start = 8)

}
}
