% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_monthly_stats.R
\name{plot_monthly_stats}
\alias{plot_monthly_stats}
\title{Plot monthly summary statistics}
\usage{
plot_monthly_stats(flowdata = NULL, HYDAT = NULL, percentiles = c(10, 20),
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, months = 1:12,
  log_discharge = FALSE, station_name = NA, write_plot = FALSE,
  write_imgtype = "pdf", write_imgsize = c(5.5, 8.5), write_dir = ".",
  na.rm = list(na.rm.global = FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{percentiles}{Numeric. Vector of percentiles to calculate. Set to NA if none required. Default \code{c(10,20)}}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{months}{Integer. Vector of months to consider for analysis (ex. \code{6:8} for Jun-Aug). Leave blank if all months
are required. Default \code{1:12}.}

\item{log_discharge}{Logical. Place the discharge axis (Y) on log scale. Default \code{TRUE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_plot}{Logical. Write the plot to specified directory. Default \code{FALSE}.}

\item{write_imgtype}{Character. One of "pdf","png","jpeg","tiff", or "bmp" image types to write the plot as. Default \code{"pdf"}.}

\item{write_imgsize}{Numeric. Height and width, respectively, of saved plot. Default \code{c(5,11)}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na.rm}{TBD}
}
\value{
Plots of summary statistics and percentiles for each month of all years
}
\description{
Plots monthly mean, median, maximum, minimum, and percentiles for each month of all years of daily flow values 
   from a streamflow dataset. Plots the statistics from all daily discharge values from all years, unless specified. Data
   calculated using the calc_monthly_stats() function.
}
\examples{
\dontrun{

plot_monthly_stats(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

plot_monthly_stats(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8, percentiles = c(1:10))

plot_monthly_stats(HYDAT = "08NM116", months = 7:9)

}
}
