% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_flow_data.R
\name{screen_flow_data}
\alias{screen_flow_data}
\title{Calculate annual summary and missing data statistics}
\usage{
screen_flow_data(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, roll_days = 1, roll_align = "right",
  water_year = FALSE, water_year_start = 10, start_year = 0,
  end_year = 9999, transpose = FALSE)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{groups}{Column in the \code{data} data frame that contains unique identifiers for different data sets. 
Only required if using the data frame option of \code{data} and groups column is not named 'STATION_NUMBER'.
Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column or identify 
another non-existing column name to remove this grouping. Identify another column if desired. Default \code{STATION_NUMBER}.}

\item{roll_days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{transpose}{Logical value indicating if the results rows and columns are to be switched. Default \code{FALSE}.}
}
\value{
A data frame with the following columns:
  \item{Year}{calendar or water year selected}
  \item{n_days}{number of days per year}
  \item{n_Q}{number of days per year with flow data}
  \item{n_missing_Q}{number of days per year with no flow data}
  \item{Minimum}{annual minimum of all daily flows for a given year}
  \item{Maximum}{annual maximum of all daily flows for a given year}
  \item{Mean}{annual mean of all daily flows for a given year}
  \item{Median}{annual median of all daily flows for a given year}
  \item{StandardDeviation}{annual 1 standard deviation of all daily flows for a given year}
  and the following monthly missing columns (order will depend on water_year_month):
  \item{Jan_missing_Q}{number of Jan days per year with no flow data}
  \item{Feb_missing_Q}{number of Feb days per year with no flow data}
  \item{Mar_missing_Q}{number of Mar days per year with no flow data}
  \item{Apr_missing_Q}{number of Apr days per year with no flow data}
  \item{May_missing_Q}{number of May days per year with no flow data}
  \item{Jun_missing_Q}{number of Jun days per year with no flow data}
  \item{Jul_missing_Q}{number of Jul days per year with no flow data}
  \item{Aug_missing_Q}{number of Aug days per year with no flow data}
  \item{Sep_missing_Q}{number of Sep days per year with no flow data}
  \item{Oct_missing_Q}{number of Oct days per year with no flow data}
  \item{Nov_missing_Q}{number of Nov days per year with no flow data}
  \item{Dec_missing_Q}{number of Dec days per year with no flow data}
  Transposing data creates a column of "Statistics" and subsequent columns for each year selected.
}
\description{
Calculates mean, median, maximum, minimum, standard deviation of annual flows and data availability and missing data
   statistics for each year and month of each year. Calculates the statistics from all daily discharge values from all years, 
   unless specified.
}
\examples{
\dontrun{

screen_flow_data(data = flow_data)

screen_flow_data(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
