% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plots.R
\name{write_plots}
\alias{write_plots}
\title{Write a plot in a directory}
\usage{
write_plots(data = NULL, file = "", digits = 10)
}
\arguments{
\item{data}{Data frame to be written to a directory.}

\item{file}{Character string naming the output file. Default filetype is .xlsx. Change to .csv using filtype argument.}

\item{digits}{Integer indicating the number of decimal places or significant digits used to round flow values. Use follows 
that of base::round() digits argument.}
}
\description{
Write plots, single or multiple, into a folder or pdf document
}
\examples{
\dontrun{

write_results(data = calc_longterm_stats(data = c("08HA002", "08HA011"),
                                         start_year = 1971, end_year = 2000), 
              file = "Cowichan River Long-term Flows (1971-2000).xlsx", 
              digits = 1)

}
}
