% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results}
\alias{write_results}
\title{Write a streamflow dataset as a .csv}
\usage{
write_results(data = NULL, station_name = "fasstr", write_dir = ".",
  write_digits = 4, na = "")
}
\arguments{
\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na}{Character. String to use for missing values in the data. Default \code{""} (blank).}

\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}
}
\value{
A .csv file of streamflow data in a selected directory.
}
\description{
Write a streamflow dataset as a .csv file to a directory. Can add missing dates or filter data by years before writing
   using given arguments. Just list flowdata data frame or HYDAT station number to write its entirety.
}
\examples{
\dontrun{

write_results(flowdata = flowdata, station_name = "MissionCreek", na="")

write_results(HYDAT = "08NM116")

}
}
