% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_basin_area.R
\name{add_basin_area}
\alias{add_basin_area}
\title{Add a basin area column to daily flows}
\usage{
add_basin_area(flow_data = NULL, flow_stations = STATION_NUMBER,
  HYDAT = NULL, basin_area = NA)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_stations}{a column in flow_data that contains station identifiers for each flow data set, if required. Default 
\code{STATION_NUMBER}.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{basin_area}{a numeric vector of basin areas. Leave blank if \code{HYDAT} is used or a column in \code{flow_data} called 
'STATION_NUMBER' contains a WSC station number, as the basin area will be extracted from HYDAT. Using \code{basin_area} will
 replace the HYDAT basin rea. If setting basin areas for multiple stations without HYDAT, set them using \code{basin_area = 
 c("08NM116" = 795, "08NM242" = 10)}; stations not listed will result in NA basin areas.}
}
\value{
A tibble data frame of the original flow_data or HYDAT data with an additional column:
  \item{Basin_Area_sqkm}{area of upstream drainage basin area, in square kilometres}
}
\description{
Add a column of basin areas to a streamflow dataset, in units of square kilometres.
}
\examples{
\dontrun{

add_basin_area(flow_data = flow_data, basin_area = 104.5)

add_basin_area(HYDAT = "08NM116")

}
}
