% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_basin_area.R
\name{add_basin_area}
\alias{add_basin_area}
\title{Add a basin area column to daily flows}
\usage{
add_basin_area(data = NULL, groups = STATION_NUMBER,
  station_number = NULL, basin_area = NA)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{basin_area}{Upstream drainage basin area, in square kilometres, to apply to observations. Three options:

(1) Leave blank if \code{groups} is STATION_NUMBER with HYDAT station numbers to extract basin areas from HYDAT.

(2) A single numeric value to apply to all observations.

(3) List each basin area for each group/station in groups (can override HYDAT value) as such \code{c("08NM116" = 795, 
"08NM242" = 10)}. Group not listed will result in NA basin area.}
}
\value{
A tibble data frame of the original source data with an additional column:
  \item{Basin_Area_sqkm}{area of upstream drainage basin area, in square kilometres}
}
\description{
Add a column of basin areas to a streamflow dataset, in units of square kilometres.
}
\examples{
\dontrun{

add_basin_area(data = flow_data, area = 104.5)

add_basin_area(station_number = "08NM116")

}
}
