% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_basin_area.R
\name{add_basin_area}
\alias{add_basin_area}
\title{Add a basin area column to daily flows}
\usage{
add_basin_area(flow_data = NULL, HYDAT = NULL, basin_area = NA)
}
\arguments{
\item{flow_data}{Data frame. A data frame of daily mean flow data. Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flow_data} argument is used.}

\item{basin_area}{Numeric. Upstream drainage basin area of the hydrometric station, in sq. km. Leave blank if \code{HYDAT} is used or 
a column in \code{flow_data} called 'STATION_NUMBER' contains a WSC station number, as the basin area will be extracted from HYDAT. 
Using \code{basin_area} will replace the HYDAT basin area. If setting basin areas for multiple stations without HYDAT, set them using 
\code{basin_area = c("08NM116" = 795, "08NM242" = 10)}; stations not listed will result in NA basin areas.}
}
\value{
A data frame of the original flow_data or HYDAT data with an additional column:
  \item{Basin_Area_sqkm}{area of upstream drainage basin area, in square kilometres}
}
\description{
Add a column of basin areas to a streamflow dataset, in units of square kilometres.
}
\examples{
\dontrun{

add_basin_area(flow_data = flow_data, basin_area = 104.5)

add_basin_area(HYDAT = "08NM116")

}
}
