% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cumulative_volume.R
\name{add_cumulative_volume}
\alias{add_cumulative_volume}
\title{Add daily cumulative volumetric flows on an annual basis}
\usage{
add_cumulative_volume(flow_data = NULL, flow_dates = Date,
  flow_values = Value, flow_stations = STATION_NUMBER, HYDAT = NULL,
  water_year = FALSE, water_year_start = 10)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_dates}{a column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Leave blank if using \code{HYDAT} 
argument. Default \code{Date}.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{flow_stations}{a column in flow_data that contains station identifiers (names or numbers) for each flow data set. Required 
if flow_data contains multiple stations. Default \code{STATION_NUMBER}.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{water_year}{a logical value indicating whether to use water years to group flow data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{a numeric value indicating the month of the start of the water year. Used if \code{water_year=TRUE}. 
Default \code{10}.}
}
\value{
A tibble data frame of the original flow_data or HYDAT data with an additional column:
  \item{Cumul_Volume_m3}{cumulative volumetric flows for each day for each year, in units of cubic metres}
}
\description{
Add a column of daily cumulative volumetric flows on an annual basis to a streamflow dataset. Adds the volumetric discharge 
   from each day with the previous day(s) for each year, in units of cubic metres. The cumulative flows restart every year.
}
\examples{
\dontrun{

add_cumulative_volume(flow_data = flow_data)

add_cumulative_volume(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
