% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cumulative_yield.R
\name{add_cumulative_yield}
\alias{add_cumulative_yield}
\title{Add daily cumulative runoff yield flows on an annual basis}
\usage{
add_cumulative_yield(flow_data = NULL, flow_dates = Date,
  flow_values = Value, flow_stations = STATION_NUMBER,
  flow_basin_areas = flow_basin_areas, HYDAT = NULL, basin_area = NA,
  water_year = FALSE, water_year_start = 10)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_dates}{a column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Leave blank if using \code{HYDAT} 
argument. Default \code{Date}.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{flow_stations}{a column in flow_data that contains station identifiers for each flow data set, if required. Default 
\code{STATION_NUMBER}.}

\item{flow_basin_areas}{a column in flow_data of numeric upstream drainage basin areas used to calculate the daily yield. If left blank
this function will use basin areas provided by the fasstr::add_basin_areas() function using the \code{basin_area} argument.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{basin_area}{a numeric vector of basin areas. If no \code{flow_basin_area} provided in flow_data, used to determine basin areas from the
fasstr::add_basin_areas() function. Leave blank if \code{HYDAT} is used or a column in \code{flow_data} called 'STATION_NUMBER' 
contains a WSC station number, as the basin area will be extracted from HYDAT. Using \code{basin_area} will replace the HYDAT basin 
area. If setting basin areas for multiple stations without HYDAT, set them using 
\code{basin_area = c("08NM116" = 795, "08NM242" = 10)}; stations not listed will result in NA basin areas.}

\item{water_year}{a logical value indicating whether to use water years to group flow data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{a numeric value indicating the month of the start of the water year. Used if \code{water_year=TRUE}. 
Default \code{10}.}
}
\value{
A tibble data frame of the original flow_data or HYDAT data with an additional column:
  \item{Cumul_Yield_mm}{cumulative yield flows for each day for each year, in units of millimetres}
}
\description{
Add a column of daily cumulative runoff yield flows on an annual basis to a streamflow dataset. Adds the runoff yield 
   discharge from each day with the previous day(s) for each year, in units of millimetres. The cumulative flows restart every year.
   Converts cumulative discharge to a depth of water based on the upstream drainge basin area.
}
\examples{
\dontrun{

add_cumulative_yield(flow_data = flow_data, basin_area = 105.6)

add_cumulative_yield(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
