% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cumulative_yield.R
\name{add_cumulative_yield}
\alias{add_cumulative_yield}
\title{Add daily cumulative runoff yield flows on an annual basis}
\usage{
add_cumulative_yield(flowdata = NULL, HYDAT = NULL, basin_area = NA,
  water_year = FALSE, water_year_start = 10)
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{basin_area}{Numeric. Upstream drainage basin area of the hydrometric station, in sq. km. Leave blank if \code{HYDAT} is used or 
a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the basin area will be extracted from HYDAT. 
Setting the basin area will replace the HYDAT basin area.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}
}
\value{
A data frame of the original flowdata or HYDAT data with an additional column:
  \item{Cumul_Yield_mm}{cumulative yield flows for each day for each year, in units of millimetres}
}
\description{
Add a column of daily cumulative runoff yield flows on an annual basis to a streamflow dataset. Adds the runoff yield 
   discharge from each day with the previous day(s) for each year, in units of millimetres. The cumulative flows restart every year.
   Converts cumulative discharge to a depth of water based on the upstream drainge basin area.
}
\examples{
\dontrun{

add_cumulative_yield(flowdata = flowdata, basin_area = 105.6)

add_cumulative_yield(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
