% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_daily_volume.R
\name{add_daily_volume}
\alias{add_daily_volume}
\title{Add daily volumetric flows}
\usage{
add_daily_volume(flow_data = NULL, flow_values = Value, HYDAT = NULL)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}
}
\value{
A tibble data frame of the original flow_data or HYDAT data with an additional column:
  \item{Volume_m3}{daily total volumetric flow, in units of cubic metres}
}
\description{
Add a column of daily volumetric flows to a streamflow dataset, in units of cubic metres. Converts the discharge to a volume.
}
\examples{
\dontrun{

add_daily_volume(flow_data = flow_data)

add_daily_volume(HYDAT = "08NM116")

}
}
