% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_variables.R
\name{add_date_variables}
\alias{add_date_variables}
\title{Add year, month, and day of year variables}
\usage{
add_date_variables(flow_data = NULL, flow_dates = Date, HYDAT = NULL,
  water_year = FALSE, water_year_start = 10)
}
\arguments{
\item{flow_data}{Data frame. A data frame of daily mean flow data. Not required if \code{HYDAT} argument is used.}

\item{flow_dates}{A column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Default \code{Date}.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flow_data} argument is used.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}
}
\value{
A data frame of the original flow_data or HYDAT data with additional columns:
  \item{Year}{calendar year}
  \item{Month}{numeric month (1 to 12)}
  \item{MonthName}{month name (Jan-Dec)}
  \item{DayofYear}{day of the year (1-365 or 366)}
  \item{WaterYear}{(optional) water year, designated by the calendar year in which it ends}
  \item{WaterDayofYear}{(optional) day of the water year (1-365 or 366), starting in the first month of the water year}
}
\description{
Add columns of Year (YYYY), Month (MM), MonthName (e.g. 'Jan'), and DayofYear (1-365 or 366); and WaterYear (YYYY) and 
   WaterDayofYear (1-365 or 366) if selected; to a data frame with a column of dates called 'Date'. Water years are designated by 
   the year in which they end. For example, Water Year 1999 (starting Oct) is from 1 Oct 1998 (WaterDayofYear 1) to 30 Sep 1999
   (WaterDayofYear 365)).
}
\examples{
\dontrun{

add_date_variables(flow_data = flow_data)

add_date_variables(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
