% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rolling_means.R
\name{add_rolling_means}
\alias{add_rolling_means}
\title{Add rolling n-day averages}
\usage{
add_rolling_means(flow_data = NULL, flow_dates = Date,
  flow_values = Value, flow_stations = STATION_NUMBER, HYDAT = NULL,
  days = c(3, 7, 30), align = "right")
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_dates}{a column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Leave blank if using \code{HYDAT} 
argument. Default \code{Date}.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{flow_stations}{a column in flow_data that contains station identifiers for each flow data set, if required. Default 
\code{STATION_NUMBER}.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{align}{a character identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{daysa}{a numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}
}
\value{
A data frame of the original flow_data or HYDAT data with additional column(s) of:
  \item{QnDay}{rolling means of the n-day flow values of the designated date and adjacent dates, direction of mean specified by align}
  Default additional columns:
  \item{Q3Day}{rolling means of the 3-day flow values of the designated date and previous 2 days (align="right")}
  \item{Q7Day}{rolling means of the 7-day flow values of the designated date and previous 6 days (align="right")}
  \item{Q30Day}{rolling means of the 30-day flow values of the designated date and previous 29 days (align="right")}
}
\description{
Adds selected n-day rolling means to a streamflow dataset. Based on selected n-days and alignment, the rolling mean for
  a given day is obtained by averaging the adjacent dates of daily mean values. For example, rolling days of '7' and 'right' alignment 
  would obtain a mean of the given and previous 6 days of daily mean flow. Rolling mean values will not be calculated if there is less 
  than the n-days provided.
}
\examples{
\dontrun{

add_rolling_means(flow_data = flow_data, days = 7, align = 'centre')

add_rolling_means(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
