% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seasons.R
\name{add_seasons}
\alias{add_seasons}
\title{Add two and four seasons identifiers}
\usage{
add_seasons(data = NULL, dates = Date, water_year = FALSE,
  water_year_start = 10)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates.
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}
}
\value{
A data frame of the source data with additional columns:
  \item{Seasons2}{two seasons identifiers}
  \item{Seasons4}{four seasons identifiers}
}
\description{
Add columns of Seasons2 (two 6-month seasons) and Seasons4 (four 3-month seasons) to a data frame with a column of dates 
   called 'Date'. The start of the two and four seasons coincides with the start month of each year ('Jan-Jun' and 'Jan-Mar' for 
   the first two and four calendar year seasons, respectively, or 'Oct-Mar' and 'Apr-Sep' for the first two and four water year
   seasons starting in Oct, respectively). Each season is designated by the calendar or water year which it occurs.
}
\examples{
\dontrun{

add_seasons(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
