% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_annual_flow_timing.R
\name{calc_annual_flow_timing}
\alias{calc_annual_flow_timing}
\title{Calculate annual flow timing}
\usage{
calc_annual_flow_timing(flowdata = NULL, HYDAT = NULL,
  percent_total = c(25, 33.3, 50, 75), water_year = FALSE,
  water_year_start = 10, start_year = NULL, end_year = NULL,
  exclude_years = NULL, transpose = FALSE, station_name = NA,
  write_table = FALSE, write_dir = ".")
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{percent_total}{Numeric. Percents of annual total flows to determine dates. Default \code{c(25,33.3,50,75)}.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{transpose}{Logical. Switch the rows and columns of the results table. Dates not transposed. Default \code{FALSE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_table}{Logical. Write the table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}
}
\value{
A data frame with the following columns:
  \item{Year}{calendar or water year selected}
  \item{DoY_'n'pct_TotalQ}{day of year for each n-percent of total volumetric discharge}
  \item{Date_'n'pct_TotalQ}{date (YYYY-MM-DD) for each n-percent of total volumetric discharge}
  Default columns:
  \item{DoY_25pct_TotalQ}{day of year of 25-percent of total volumetric discharge}
  \item{Date_25pct_TotalQ}{date (YYYY-MM-DD) of 25-percent of total volumetric discharge}
  \item{DoY_33.3pct_TotalQ}{day of year of 33.3-percent of total volumetric discharge}
  \item{Date_33.3pct_TotalQ}{date (YYYY-MM-DD) of 33.3-percent of total volumetric discharge}
  \item{DoY_50pct_TotalQ}{day of year of 50-percent of total volumetric discharge}
  \item{Date_50pct_TotalQ}{date (YYYY-MM-DD) of 50-percent of total volumetric discharge}
  \item{DoY_75pct_TotalQ}{day of year of 75-percent of total volumetric discharge}
  \item{Date_75pct_TotalQ}{date (YYYY-MM-DD) of 75-percent of total volumetric discharge}
  Transposing data creates a column of "Statistics" and subsequent columns for each year selected.
}
\description{
Calculates annual the timing (day of year) and date of occurrence of portions of total annual flow of daily flow 
   values from a streamflow dataset. Calculates the statistics from all daily discharge values from all years, unless specified.
}
\examples{
\dontrun{

calc_annual_flow_timing(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

calc_annual_flow_timing(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8, percent_total = 50)

}
}
