% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lt_mad.R
\name{calc_lt_mad}
\alias{calc_lt_mad}
\title{Calculate the long-term mean annual discharge}
\usage{
calc_lt_mad(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, roll_days = 1, roll_align = "right",
  water_year = FALSE, water_year_start = 10, start_year = 0,
  end_year = 9999, exclude_years = NULL, complete_years = FALSE,
  months = 1:12, percent_MAD = NA, transpose = FALSE)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{groups}{Column in the \code{data} data frame that contains unique identifiers for different data sets. 
Only required if using the data frame option of \code{data} and groups column is not named 'STATION_NUMBER'.
Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column or identify 
another non-existing column name to remove this grouping. Identify another column if desired. Default \code{STATION_NUMBER}.}

\item{roll_days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{complete_years}{Logical values indicating whether to include only years with complete data in analysis. Default \code{FALSE}.}

\item{months}{Numeric vector of months to include in analysis (ex. \code{6:8} for Jun-Aug). Leave blank to summarize 
all months (default \code{1:12}).}

\item{percent_MAD}{Numeric vector of percents of long-term mean annual discharge to add to the table (ex. 20 for 20 percent MAD).
Leave blank for no values to be calculated.}

\item{transpose}{Logical value indicating if the results rows and columns are to be switched. Default \code{FALSE}.}
}
\value{
A tibble of numeric values of a long-term mean (and percent of long-term mean if selected) of selected years and months.
}
\description{
Calculates the long-term mean annual discharge of a streamflow dataset. Averages all daily discharge values from all years,
  unless specified.
}
\examples{
\dontrun{

calc_lt_mad(data = data, start_year = 1980, end_year = 2010)

calc_lt_mad(data = "08NM116", water_year = TRUE, exclude_years = (1990, 1992:1994))

calc_lt_mad(data = "08NM116", percent_MAD = 20)

}
}
