% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_monthly_stats.R
\name{calc_monthly_stats}
\alias{calc_monthly_stats}
\title{Calculate monthly summary statistics}
\usage{
calc_monthly_stats(flowdata = NULL, HYDAT = NULL, percentiles = c(10, 20),
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, months = 1:12, spread = FALSE,
  transpose = FALSE, station_name = NA, write_table = FALSE,
  write_digits = 3, write_dir = ".", na.rm = list(na.rm.global = FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{percentiles}{Numeric. Vector of percentiles to calculate. Set to NA if none required. Default \code{c(10,20)}}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{months}{Integer. Vector of months to consider for analysis (ex. \code{6:8} for Jun-Aug). Leave blank if all months
are required. Default \code{1:12}.}

\item{spread}{Logical. All monthly statistics as column names. Default \code{FALSE}.}

\item{transpose}{Logical. All monthly statistics as row names. Default \code{FALSE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_table}{Logical. Write the table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_digits}{Numeric. Number of significant digits to round the results in the written table. Default \code{3}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na.rm}{TBD}
}
\value{
A data frame with the following columns:
  \item{Year}{calendar or water year selected}
  \item{Month}{month of the year}
  \item{Mean}{mean of all daily flows for a given month and year}
  \item{Median}{median of all daily flows for a given month and year}
  \item{Maximum}{maximum of all daily flows for a given month and year}
  \item{Minimum}{minimum of all daily flows for a given month and year}
  \item{P'n'}{each n-th percentile selected for a given month and year}
  Default percentile columns:
  \item{P10}{10th percentile of all daily flows for a given month and year}
  \item{P20}{20th percentile of all daily flows for a given month and year}
  Transposing data creates a column of "Statistics" for each month, labeled as "Month-Statistic" (ex "Jan-Mean"),
  and subsequent columns for each year selected.
  Spreading data creates columns of Year and subsequent columns of Month-Statistics  (ex "Jan-Mean").
}
\description{
Calculates monthly mean, median, maximum, minimum, and percentiles for each month of all years of daily flow values 
   from a streamflow dataset. Calculates the statistics from all daily discharge values from all years, unless specified.
}
\examples{
\dontrun{

calc_monthly_stats(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

calc_monthly_stats(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8, percentiles = c(1:10))

calc_monthly_stats(HYDAT = "08NM116", months = 7:9)

}
}
