% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_annual_trends.R
\name{compute_annual_trends}
\alias{compute_annual_trends}
\title{Calculate prewhitened nonlinear trends on streamflow data}
\usage{
compute_annual_trends(flowdata = NULL, HYDAT = NULL, trendsdata = NULL,
  zyp_method = NA, basin_area = NA, water_year = FALSE,
  water_year_start = 10, start_year = NULL, end_year = NULL,
  exclude_years = NULL, annual_percentiles = c(10, 90),
  monthly_percentiles = c(10, 20), lowflow_days = c(1, 3, 7, 30),
  lowflow_align = "right", totalflow_seasons = TRUE,
  timing_percent = c(25, 33, 50, 75), normal_percentiles = c(25, 75),
  transpose = FALSE, station_name = NA, write_table_data = FALSE,
  write_table_results = FALSE, write_dir = ".", na.rm = list(na.rm.global
  = FALSE))
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{trendsdata}{Data frame. Annual data with column names of years and rows of annual statistics. Leave blank if using \code{HYDAT}}

\item{zyp_method}{Character. The prewhitened trend method to use from 'zyp', either "zhang' or "yuepilon". Required.}

\item{basin_area}{Numeric. Upstream drainage basin area of the hydrometric station, in sq. km. Leave blank if \code{HYDAT} is used or 
a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the basin area will be extracted from HYDAT. 
Setting the basin area will replace the HYDAT basin area.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{Integer. Single year or vector of years to exclude from analysis. Leave blank if all years are required.}

\item{annual_percentiles}{Numeric. Vector of percentiles to calculate annually. Set to NA if none required. Used for calc_annual_stats()
function. Default \code{c(10,90)}.}

\item{monthly_percentiles}{Numeric. Vector of percentiles to calculate monthly for each year. Set to NA if none required. Used for 
calc_monthly_stats() function. Default \code{c(10,90)}.}

\item{lowflow_days}{Numeric. The number of days to apply a rolling mean. Used for calc_annual_lowflows() function. Default \code{1}.}

\item{lowflow_align}{Character. Specifies whether the dates of the rolling mean should be specified by the first ('left'), last ('right),
or middle ('center') of the rolling n-day group of observations. Used for calc_annual_lowflows() function. Default \code{'right'}.}

\item{totalflow_seasons}{Logical. Include seasonal yields and total discharges.Used for calc_annual_total_flows() function. 
Default \code{TRUE}.}

\item{timing_percent}{Numeric. Percents of annual total flows to determine dates. Used for calc_annual_flow_timing() function. 
Default \code{c(25,33.3,50,75)}.}

\item{normal_percentiles}{Numeric. Lower and upper percentiles, respectively indicating the limits of the normal range. 
Used for calc_annual_outside_normal() function. Default \code{c(25,75)}.}

\item{transpose}{Logical. Switch the rows and columns of the results table. Default \code{FALSE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_table_data}{Logical. Write the trending input data table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_table_results}{Logical. Write the trending results table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}

\item{na.rm}{TBD}
}
\value{
A data frame containing trends with the following outputs from the zyp package:
  \item{Statistic}{the annual statistic used for trending.}
  \item{lbound}{the lower bound of the trend's confidence interval}
  \item{trend}{the Sens' slope (trend) per unit time}
  \item{trendp}{the Sen's slope (trend) over the time period}
  \item{ubound}{the upper bound of the trend's confidence interval}
  \item{tau}{Kendall's tau statistic computed on the final detrended timeseries}
  \item{sig}{Kendall's P-value computed for the final detrended timeseries}
  \item{nruns}{the number of runs required to converge upon a trend}
  \item{autocor}{the autocorrelation of the final detrended timeseries}
  \item{valid_frac}{the fraction of the data which is valid (not NA) once autocorrelation is removed}
  \item{linear}{the least squares fit trend on the same dat}
  \item{intercept}{the lower bound of the trend's confidence interval}
  \item{lbound}{the intercept of the Sen's slope (trend)}
  and the following additional columns:
  \item{min_year}{the minimum year used in the trending}
  \item{max_year}{the maximum year used in the trending}
  \item{n_years}{the number of years with data for trending}
  \item{mean}{the mean of all values used for trending}
  \item{median}{the median of all values used for trending}
  \item{min}{the minimum of all values used for trending}
  \item{max}{the maximum of all values used for trending}
  Transposing data creates a column of "Trends_Statistics" and subsequent columns of each trends statistic for each annual statistic.
}
\description{
Calculates prewhitened nonlinear trends on annual streamflow data. Uses the
   \href{https://cran.r-project.org/web/packages/zyp/index.html}{'zyp'} package to trend. Review 'zyp' to understand its methology.
   Use \code{trendsdata} if providing your own data frame of statistics to trend. If using \code{HYDAT} or \code{flowdata}, then all
   annual statistics will be calculated using the calc_all_annual_stats() function which uses the following fasstr functions:
\itemize{
 \item{calc_annual_stats()}
 \item{calc_annual_lowflows()}
 \item{calc_annual_total_flows()}
 \item{calc_annual_flow_timing()}
 \item{calc_monthly_stats()}
 \item{calc_annual_outside_normal()}
 }
}
\examples{
\dontrun{

compute_annual_trends(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

compute_annual_trends(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
