% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr-package.R
\docType{package}
\name{fasstr-package}
\alias{fasstr}
\alias{fasstr-package}
\title{fasstr: Tools for Streamflow Data Tidying, Analyses, and Visualization}
\description{
'fasstr' (Flow Analysis Summary Statistics Tool for R) provides useful functions to tidy, 
    analyze, and visualize daily streamflow time series data. This package provides functions with quick
    solutions for: screening for missing dates and outliers; summarizing and plotting various daily, month, 
    and long-term statistics; trending annual patterns; and computing low and high-flow frequency analyses. 
    Various function arguments allow for easy customization of analyses and plotting. While not required
    for use, this package integrates the 'tidyhydat' package to extract and tidy streamflow data from a 
    Water Survey of Canada HYDAT database.
}
\references{
tidyhydat information:


To use the station_number argument of the fasstr functions, please download the latest version of HYDAT using the function:
 \itemize{
  \item \code{tidyhydat::download_hydat()}
  }


 For more information on HYDAT 
 \itemize{
   \item Please see this description of the database: goo.gl/H3NXJQ
   \item This page is landing page for technical description of HYDAT:
   \url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/}
   \item This page links to a document that outlines database table definitions:
   \url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/HYDAT_Definition_EN.pdf}
   }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/fasstr}
  \item Report bugs at \url{https://github.com/bcgov/fasstr/issues}
}

}
\author{
\strong{Maintainer}: Jon Goetz \email{jon.goetz@gov.bc.ca}

Authors:
\itemize{
  \item Carl James Schwarz \email{cschwarz.stat.sfu.ca@gmail.com}
}

Other contributors:
\itemize{
  \item Robin Pike \email{robin.g.pike@gov.bc.ca} [contributor]
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
