% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_annual_freq_analysis.R
\name{fasstr_annual_freq_analysis}
\alias{fasstr_annual_freq_analysis}
\title{Perform a volume frequency analysis on annual statistics.}
\usage{
fasstr_annual_freq_analysis(station_name = "fasstr", flowdata = NULL,
  HYDAT = NULL, HYDAT_peaks = NA, water_year = FALSE, start_year = NULL,
  end_year = NULL, exclude_years = NULL, rolling_days = c(3, 7, 30),
  use_log = FALSE, use_max = FALSE, prob_plot_position = c("weibull",
  "median", "hazen"), prob_scale_points = c(0.9999, 0.999, 0.99, 0.9, 0.5,
  0.2, 0.1, 0.02, 0.01, 0.001, 1e-04), fit_distr = c("PIII", "weibull"),
  fit_distr_method = ifelse(fit_distr == "PIII", "MOM", "MLE"),
  fit_quantiles = c(0.975, 0.99, 0.98, 0.95, 0.9, 0.8, 0.5, 0.2, 0.1, 0.05,
  0.01), na.rm = list(na.rm.global = TRUE), write_stat_table = TRUE,
  write_plotdata_table = TRUE, write_quantiles_table = TRUE,
  write_frequency_plot = TRUE, write_frequency_plot_type = c("pdf", "png"),
  report_dir = ".", table_nddigits = 3)
}
\arguments{
\item{station_name}{Character. Identifier name of the stream or station. Required when supplying data through \code{flowdata}.
The station name will be used in plots and filenames of exported tables and plot. If using \code{HYDAT} to supply
data and no \code{station_name} is provided, the HYDAT station number will be the identifier.}

\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics.
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Value' column with the daily
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{HYDAT_peaks}{Character. "MAX" or "MIN" instantaneous peaks pulled from HYDAT}

\item{water_year}{Logical. Set to \code{TRUE} if data should be summarized by water year (Oct-Sep) instead of the
default calendar year (Jan-Dec) (\code{water_year=FALSE}). Water years are designated by the year which they end in
(e.g. water year 2000 start on 1 Oct 1999 and ends on 30 Sep 2000).}

\item{start_year}{Numeric. The first year of streamflow data to analyze. If unset, the default \code{start_year} is the first
year of the data provided.}

\item{end_year}{Numeric. The last year of streamflow data to analyze. If unset, the default \code{end_year} is the last
year of the data provided.}

\item{exclude_years}{Numeric. List of years to exclude final results from. Ex. 1990 or c(1990,1995:2000).}

\item{rolling_days}{Volume frequency analysis conducted on these rolling averages.}

\item{use_log}{Transfrom to log-scale before analysis?}

\item{use_max}{Analyze the maximums rather than the minimums.}

\item{prob_plot_position}{Which plotting positions should be used in the frequency plots. Points are plotted
against  (i-a)/(n+1-a-b) where \code{i} is the rank of the value; \code{n} is the sample size and
\code{a} and \code{b} are defined as:
(a=0, b=0) for Weibull plotting positions;
(a=.2; b=.3) for Median plotting postions;
(a=.5; b=.5) for Hazen plotting positions.}

\item{prob_scale_points}{What points should be plotted along the $X$ axis in the frequency plot.}

\item{fit_distr}{Which distribution should be fit? PIII = Pearson Log III distribution; weibull=Weibull distribution.}

\item{fit_distr_method}{Which method used to fit the distribution. MOM=Method of moments; MLE=maximum likelihood estimation.}

\item{fit_quantiles}{Which quantiles should be estimated from the fitted distribution?}

\item{na.rm}{}

\item{write_stat_table}{Should a file be created with the computed percentiles?
The file name will be  \code{file.path(report_dir,paste(station_name,"-annual-vfa-stat.csv", sep=""))}.}

\item{write_plotdata_table}{Should a file be created with the frequency plot data?
The file name will be  \code{file.path(report_dir, paste(station_name,"-annual-vfa-plotdata.csv", sep=""))}.}

\item{write_quantiles_table}{Should a file be created with the fitted quantiles?.
The file name will be  \code{file.path(report_dir, paste(station_name,"-annual-vfa-quantiles.csv", sep=""))}.}

\item{write_frequency_plot}{Should a file be created with the frequency plot..
The file name will be \code{file.path(report_dir, paste(station_name,"-annual-vfa-frequency-plot.",write_frequency_plot_type[1],sep=""))}}

\item{write_frequency_plot_type}{Format of the frequency plot.}

\item{report_dir}{Character. Folder location of where to write tables and plots. Default is the working directory.}

\item{table_nddigits}{Numeric. Number of significant digits to round the results in the written tables. Default is 3.}
}
\value{
A list with the following elements:
  \item{Q.flow.summary}{Data frame with flow summary.}
  \item{start_year}{Start year of the analysis}
  \item{end_year}{End year of the analysis}
  \item{water_year}{Were computations done on water year?}
  \item{use_max}{Were computations done on maximum values.}
  \item{rolling_days}{Rolling average days on which statistics computed.}
  \item{Q_stat}{Data frame with Computed annual summary statistics used in analysis}
  \item{Q_stat_trans}{Data frame with Computed annual summary statistics (transposed) used in analysis.}
  \item{plotdata}{Data frame with Co-ordinates used in frequency plot.}
  \item{prob_plot_position}{Which plotting position was used in frequency plot?}
  \item{freqplot}{ggplot2 object with frequency plot}
  \item{fit_distr}{Which distribution was fit to the data}
  \item{fit}{List of fitted objects from fitdistrplus.}
  \item{fitted_quantiles}{Data frame with fitted quantiles.}
  \item{fitted_quantiles_trans}{Data frame with fitted quantiles (transposed)}
  \item{file_stat_csv}{Object with file name of *.csv file with flow summary}
  \item{ file_stat_trans_csv}{Object with file name of *.csv file with flow summary (transposed)}
  \item{file_plotdata_csv}{Object with file name of *.csv file with plotting information for frequency plot.}
  \item{file_quantile_csv}{Object with file name of fitted quantiles.}
  \item{file_quantile_trans_csv}{Object with file name of fitted quantiles (transposed)}
  \item{file_frequency_plot}{Object with file name of *.pdf or *.png file with frequency plot}
  \item{Version}{Version of this function.}
  \item{Date}{Date function was run.}
}
\description{
Performs a volume frequency analysis on annual statistics similar to HEC-SSP.
}
\examples{
\dontrun{
vfa.analysis <- compute.volume.frequency.analysis(
                     station_name ='XXX',
                     flowdata         =flow,
                     start_year   =1960,
                     end_year     =2014)
}
}
