% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_annual_missing_plots.R
\name{fasstr_annual_missing_plots}
\alias{fasstr_annual_missing_plots}
\title{Plot annual and monthly missing dates}
\usage{
fasstr_annual_missing_plots(flowdata = NULL, HYDAT = NULL,
  rolling_days = 1, rolling_align = "right", water_year = FALSE,
  water_year_start = 10, start_year = NULL, end_year = NULL,
  station_name = NA, write_plot = FALSE, write_imgtype = "pdf",
  write_imgsize = c(6, 8.5), write_dir = ".")
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{rolling_days}{Numeric. The number of days to apply a rolling mean. Default \code{1}.}

\item{rolling_align}{Character. Specifies whether the dates of the rolling mean should be specified by the first ('left'), last ('right),
or middle ('center') of the rolling n-day group of observations. Default \code{'right'}.
by the year in which they end. Default \code{FALSE}.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_plot}{Logical. Write the plot to specified directory. Default \code{FALSE}.}

\item{write_imgtype}{Character. One of "pdf","png","jpeg","tiff", or "bmp" image types to write the plot as. Default \code{"pdf"}.}

\item{write_imgsize}{Numeric. Height and width, respectively, of saved plot. Default \code{c(6,8.5)}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}
}
\value{
A ggplot2 object with plots for each month counting the number of missing dates for each year
}
\description{
Plots the number of missing data for each month of each year. Calculates the statistics from all daily discharge 
   values from all years, unless specified. Data calculated using fasstr_data_screening() function.
}
\examples{
\dontrun{

fasstr_annual_missing_plots(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

fasstr_annual_missing_plots(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
