% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_annual_trends_analysis.R
\name{fasstr_annual_trends_analysis}
\alias{fasstr_annual_trends_analysis}
\title{Compute annual trends.}
\usage{
fasstr_annual_trends_analysis(trendsdata = NULL, flowdata = NULL,
  HYDAT = NULL, zyp_method = NA, station_name = "fasstr",
  water_year = FALSE, water_year_start = 10, start_year = NULL,
  end_year = NULL, exclude_years = NULL, basin_area = NA,
  write_trends_data = FALSE, write_trends_results = FALSE,
  report_dir = ".", na.rm = list(na.rm.global = FALSE),
  table_nddigits = 3)
}
\arguments{
\item{trendsdata}{Dataframe. Annual data with column names of years and rows of annual statistics.}

\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Value' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{station_name}{Character. Identifier name of the stream or station. Required when supplying data through \code{flowdata}.
The station name will be used in plots and filenames of exported tables and plot. If using \code{HYDAT} to supply
data and no \code{station_name} is provided, the HYDAT station number will be the identifier.}

\item{water_year}{Logical. Set to \code{TRUE} if data should be summarized by water year (Oct-Sep) instead of the
default calendar year (Jan-Dec) (\code{water_year=FALSE}). Water years are designated by the year which they end in
(e.g. water year 2000 start on 1 Oct 1999 and ends on 30 Sep 2000).}

\item{start_year}{Numeric. The first year of streamflow data to analyze. If unset, the default \code{start_year} is the first
year of the data provided.}

\item{end_year}{Numeric. The last year of streamflow data to analyze. If unset, the default \code{end_year} is the last
year of the data provided.}

\item{basin_area}{Numeric. The upstream drainage basin area (in sq. km) of the station. Used to calculate runoff yields (mm).
If no value provided, yield calculations will result in NA values.}

\item{report_dir}{Character. Folder location of where to write tables and plots. Default is the working directory.}

\item{na.rm}{TBD}

\item{table_nddigits}{Numeric. Number of significant digits to round the results in the written tables. Default is 3.}

\item{excluded.years}{Numeric. List of years to exclude final results from. Ex. 1990 or c(1990,1995:2000).}

\item{write_table}{Logical. Should a file be created with the calendar year computed percentiles?
The file name will be  \code{file.path(report_dir,paste(station_name,'-annual-cy-summary-stat.csv'))}.}

\item{write_transposed_table}{Logical. Should a file be created with the transposed of the annual statistics
(both calendar and water year)?
The file name will be  \code{file.path(report_dir,paste(station_name,'-annual-summary-stat-trans.csv'))}.}
}
\description{
Computes annual trends.
}
\examples{
\dontrun{

Coming

}
}
