% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstr_LTMAD.R
\name{fasstr_LTMAD}
\alias{fasstr_LTMAD}
\title{Compute long-term mean annual discharge.}
\usage{
fasstr_LTMAD(flowdata = NULL, HYDAT = NULL, water_year = FALSE,
  water_year_start = 10, start_year = NULL, end_year = NULL,
  exclude_years = NULL, percent_MAD = 100)
}
\arguments{
\item{flowdata}{Dataframe. A dataframe of daily mean streamflow data used to calculate the annual statistics. 
Two columns are required: a 'Date' column with dates formatted YYYY-MM-DD and a 'Value' column with the daily 
mean streamflow values in units of cubic metres per second. \code{flowdata} not required if \code{HYDAT} is used.}

\item{HYDAT}{Character. A HYDAT station number (e.g. "08NM116") of which to extract daily streamflow data from the HYDAT database.
tidyhydat package and a downloaded SQLite HYDAT required.}

\item{water_year}{Logical. Set to \code{TRUE} if data should be summarized by water year (Oct-Sep) instead of the
default calendar year (Jan-Dec) (\code{water_year=FALSE}). Water years are designated by the year which they end in
(e.g. water year 2000 start on 1 Oct 1999 and ends on 30 Sep 2000).}

\item{water_year_start}{Numeric. Month to start water year (1 to 12 for Jan to Dec).}

\item{start_year}{Numeric. The first year of streamflow data to analyze. If unset, the default \code{start_year} is the first
year of the data provided.}

\item{end_year}{Numeric. The last year of streamflow data to analyze. If unset, the default \code{end_year} is the last
year of the data provided.}

\item{percent_MAD}{Numeric. Factor to mulitply the MAD. Deafult 100 (i.e. 100%MAD)}

\item{excluded_years}{Numeric. List of years to exclude final results from. Ex. 1990 or c(1990,1995:2000).}
}
\description{
Computes long-term statistics of streamflow data.
Streamflow data can be supplied through the \code{flowdata} parameter or extracted from a 
HYDAT database using the tidyhydat package and \code{HYDAT} parameter.
}
\examples{
\dontrun{

coming soon :)
}
}
