% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_dates.R
\name{fill_missing_dates}
\alias{fill_missing_dates}
\title{Fill dates of missing flow values with NA}
\usage{
fill_missing_dates(flow_data = NULL, flow_dates = Date,
  flow_values = Value, HYDAT = NULL, water_year = FALSE,
  water_year_start = 10)
}
\arguments{
\item{flow_data}{Data frame. A data frame of daily mean flow data. Not required if \code{HYDAT} argument is used.}

\item{flow_dates}{A column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Default \code{Date}.}

\item{flow_values}{A column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Default \code{Value}.}

\item{HYDAT}{Character. Seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flow_data} argument is used.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}
}
\value{
A data frame of the original flow_data or HYDAT data with additional rows of filled values of missing dates with NA
}
\description{
Adds rows of dates with missing flow values to a streamflow dataset with daily flow values of NA. Missing dates will 
   be filled in gaps between data and compeltely fill the first and last years (calendar or water year if selected).
}
\examples{
\dontrun{

fill_missing_dates(flow_data = flow_data)

fill_missing_dates(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
