% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_dates.R
\name{fill_missing_dates}
\alias{fill_missing_dates}
\title{Fill dates of missing flow values with NA}
\usage{
fill_missing_dates(data, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number, water_year_start = 1)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year for analysis. Default \code{1}.}
}
\value{
A tibble data frame of the source data with additional rows of filled values of missing dates.
}
\description{
Adds rows of dates with missing flow values to a streamflow dataset with daily flow values of NA. Missing dates will 
   be filled in gaps between data and compeltely fill the first and last years (calendar or water year if selected).
}
\examples{
\dontrun{

# Fill missing dates with NA using calendar years
fill_missing_dates(data = "08NM116")

# Fill missing dates with NA using water years starting in August
fill_missing_dates(data = "08NM116", 
                   water_year_start = 8)
}
}
