% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_cumulative_stats.R
\name{plot_annual_cumulative_stats}
\alias{plot_annual_cumulative_stats}
\title{Plot annual and seasonal total flows}
\usage{
plot_annual_cumulative_stats(data = NULL, dates = Date, values = Value,
  use_yield = FALSE, basin_area = NA, water_year = FALSE,
  water_year_start = 10, start_year = 0, end_year = 9999,
  exclude_years = NULL, incl_seasons = FALSE, log_discharge = FALSE)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates and values.
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{use_yield}{Logical value indicating whether to use yield runoff, in mm, instead of volumetric. Default \code{FALSE}.}

\item{basin_area}{Upstream drainage basin area to apply to daily observations. Options:

Leave blank if \code{groups} is STATION_NUMBER with HYDAT station numbers to extract basin areas from HYDAT.

Single numeric value to apply to all observations.

List each basin area for each grouping factor (can override HYDAT value) as such \code{c("08NM116" = 795, "08NM242" = 10)}.
Factors not listed will result in NA basin areas.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{incl_seasons}{Logical value indication whether to include seasonal yields and total discharges. Default \code{TRUE}.}

\item{log_discharge}{Logical value to indicate plotting the discharge axis (Y-axis) on a logarithmic scale. Default \code{TRUE}.}
}
\value{
A list of the following ggplot2 objects (yield plots unavailable if no basin_area):
  \item{TotalQ_Annual}{ggplot2 object of annual total volumetric discharge, in cubic metres}
  \item{TotalQ_Two_Seasons}{ggplot2 object of Oct-Mar and Apr-Sep total volumetric discharges, in cubic metres}
  \item{TotalQ_Four_Seasons}{ggplot2 object of Jan-Mar, Apr-Jun, Jul-Sep, Oct-Dec total volumetric discharges, in cubic metres}
  \item{Yield_Annual}{ggplot2 object of annual runoff yield, in millimetres}
  \item{Yield_Two_Seasons}{ggplot2 object of Oct-Mar and Apr-Sep runoff yields, in millimetres}
  \item{Yield_Four_Seasons}{ggplot2 object of Jan-Mar, Apr-Jun, Jul-Sep, Oct-Dec runoff yields, in millimetres}
}
\description{
Plots annual and seasonal total flows, volumetric and runoff yield flows, from a streamflow dataset. Calculates 
   the statistics from all daily discharge values from all years, unless specified. Data calculated from
   plot_annual_cumulative_stats() function. For water year and seasonal data, the designated
   year is the year in which the year or season ends. For example, if using water years with a start month of 11, the OND season is
   designated by the water year which starts in November (designated by the calendar year in which it ends).
}
\examples{
\dontrun{


plot_annual_cumulative_stats(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
