% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_lowflows.R
\name{plot_annual_lowflows}
\alias{plot_annual_lowflows}
\title{Plot annual lowflows}
\usage{
plot_annual_lowflows(data, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number, roll_days = c(1, 3, 7, 30),
  roll_align = "right", water_year_start = 1, start_year, end_year,
  exclude_years, months = 1:12, ignore_missing = FALSE,
  log_discharge = FALSE, include_title = FALSE)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{roll_days}{Numeric value of the number of days to apply a rolling mean. Default \code{1}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first 
('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year for analysis. Default \code{1}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{months}{Numeric vector of months to include in analysis (e.g.. \code{6:8} for Jun-Aug). Leave blank to summarize 
all months (default \code{1:12}).}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{FALSE}.}

\item{log_discharge}{Logical value to indicate plotting the discharge axis (Y-axis) on a logarithmic scale. Default \code{FALSE}.}

\item{include_title}{Logical value to indicate adding the group/station number to the plot, if provided. Default \code{FALSE}.}
}
\value{
A list of ggplot2 objects with the following for each station provided:
  \item{Annual_Minimums}{ggplot2 object of annual minimums of selected n-day rolling means}
  \item{Annual_Minimums_Days}{ggplot2 object of the day of years of annual minimums of selected n-day rolling means}
}
\description{
Plot annual n-day minimum values, and the day of year of occurrence of daily flow values from a 
   streamflow dataset. Plots statistics from all daily discharge values from all years, unless specified. Data
   calculated from calc_annual_lowflows() function.
}
\examples{
\dontrun{

# Plot statistics with default rolling days and alignment
plot_annual_lowflows(station_number = "08NM116") 

# Plot statistics with custom rolling days and alignment
plot_annual_lowflows(station_number = "08NM116",
                     roll_days = c(3,7),
                     roll_align = "center")
}
}
\seealso{
\code{\link{calc_annual_lowflows}}
}
