% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily_cumulative_stats.R
\name{plot_daily_cumulative_stats}
\alias{plot_daily_cumulative_stats}
\title{Plot cumulative daily flow statistics}
\usage{
plot_daily_cumulative_stats(data, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number, use_yield = FALSE,
  basin_area, water_year_start = 1, start_year, end_year, exclude_years,
  log_discharge = FALSE, include_title = FALSE, include_year)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{use_yield}{Logical value indicating whether to use yield runoff, in mm, instead of volumetric. Default \code{FALSE}.}

\item{basin_area}{Upstream drainage basin area, in square kilometres, to apply to observations. Three options:

(1) Leave blank if \code{groups} is STATION_NUMBER with HYDAT station numbers to extract basin areas from HYDAT.

(2) A single numeric value to apply to all observations.

(3) List each basin area for each group/station in groups (can override HYDAT value if listed) as such \code{c("08NM116" = 795, 
"08NM242" = 10)}. If group is not listed the HYDAT area will be applied if it exists, otherwise it will be NA.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year for analysis. Default \code{1}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{log_discharge}{Logical value to indicate plotting the discharge axis (Y-axis) on a logarithmic scale. Default \code{FALSE}.}

\item{include_title}{Logical value to indicate adding the group/station number to the plot, if provided. Default \code{FALSE}.}

\item{include_year}{A numeric value indicating a year of daily flows to add to the daily statistics plot. Leave blank for no years.}
}
\value{
A list of ggplot2 objects with the following for each station provided:
  \item{Daily_Cumulative_Stats}{a plot that contains daily cumulative flow statistics}
  Default plots on each object:   
  \item{Mean}{daily cumulative mean}
  \item{Median}{daily cumulative median}
  \item{Min-5 Percentile Range}{a ribbon showing the range of data between the daily cumulative minimum and 5th percentile}
  \item{5-25 Percentiles Range}{a ribbon showing the range of data between the daily cumulative 5th and 25th percentiles}
  \item{25-75 Percentiles Range}{a ribbon showing the range of data between the daily cumulative 25th and 75th percentiles}
  \item{75-95 Percentiles Range}{a ribbon showing the range of data between the daily cumulative 75th and 95th percentiles}
  \item{95 Percentile-Max Range}{a ribbon showing the range of data between the daily cumulative 95th percentile and the maximum}
  \item{'Year' Flows}{(optional) the daily cumulative flows for the designated year}
}
\description{
Plot the daily cumulative mean, median, maximum, minimum, and 5, 25, 75, 95th percentiles for each day of the year 
   from a streamflow dataset. Plots the statistics from all daily cumulative values from all years, unless specified. 
   Data calculated using calc_daily_cumulative_stats() function. Can plot individual years for comparision using the 
   include_year argument. Defaults to volumetric cumulative flows, can use \code{use_yield} and \code{basin_area} to convert to 
   runoff yield.
}
\examples{
\dontrun{

# Plot volume statistics
plot_daily_cumulative_stats(station_number = "08NM116") 

# Plot yield statistics with default HYDAT basin area
plot_daily_cumulative_stats(station_number = "08NM116",
                            use_yield = TRUE) 

# Plot yield statistics with custom basin area
plot_daily_cumulative_stats(station_number = "08NM116",
                            use_yield = TRUE,
                            basin_area = 800) 
}
}
\seealso{
\code{\link{calc_daily_cumulative_stats}}
}
