% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily_stats.R
\name{plot_daily_stats}
\alias{plot_daily_stats}
\title{Plot daily summary statistics}
\usage{
plot_daily_stats(data, dates = Date, values = Value,
  groups = STATION_NUMBER, station_number, roll_days = 1,
  roll_align = "right", water_year_start = 1, start_year, end_year,
  exclude_years, complete_years = FALSE, months = 1:12,
  ignore_missing = FALSE, log_discharge = TRUE,
  include_title = FALSE, include_year)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{groups}{Column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required if
groups column name is not 'STATION_NUMBER'. Function will automatically group by a column named 'STATION_NUMBER' if present.
Remove the 'STATION_NUMBER' column beforehand to remove this grouping. Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{roll_days}{Numeric value of the number of days to apply a rolling mean. Default \code{1}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first 
('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year for analysis. Default \code{1}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{complete_years}{Logical values indicating whether to include only years with complete data in analysis. Default \code{FALSE}.}

\item{months}{Numeric vector of months to include in analysis (e.g.. \code{6:8} for Jun-Aug). Leave blank to summarize 
all months (default \code{1:12}).}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{FALSE}.}

\item{log_discharge}{Logical value to indicate plotting the discharge axis (Y-axis) on a logarithmic scale. Default \code{FALSE}.}

\item{include_title}{Logical value to indicate adding the group/station number to the plot, if provided. Default \code{FALSE}.}

\item{include_year}{A numeric value indicating a year of daily flows to add to the daily statistics plot. Leave blank for no years.}
}
\value{
A list of ggplot2 objects with the following for each station provided:
  \item{Daily_Stats}{a plot that contains daily flow statistics}
  Default plots on each object:  
  \item{Mean}{daily mean}
  \item{Median}{daily median}
  \item{25-75 Percentiles Range}{a ribbon showing the range of data between the daily 25th and 75th percentiles}
  \item{5-95 Percentiles Range}{a ribbon showing the range of data between the daily 5th and 95th percentiles}
  \item{Max-Min Range}{a ribbon showing the range of data between the daily minimum and maximums}
  \item{'Year' Flows}{(on annual plots) the daily flows for the designated year}
}
\description{
Plot the daily mean, median, maximum, minimum, and percentiles for each day of the year of daily flow values 
   from a streamflow dataset. Plots the statistics from all daily discharge values from all years, unless specified. Can determine
   statistics of rolling mean days (e.g. 7-day flows) using the roll_days argument. Data calculated using calc_daily_stats()
   function.
}
\examples{
\dontrun{

# Plot statistics using data argument with defaults
flow_data <- tidyhydat::hy_daily_flows(station_number = "08NM116")
plot_daily_stats(data = flow_data,
                 start_year = 1980)

# Plot statistics using station_number argument with defaults
plot_daily_stats(station_number = "08NM116",
                 start_year = 1980)

# Plot statistics regardless if there is missing data for a given day of year
plot_daily_stats(station_number = "08NM116",
                 ignore_missing = TRUE)
                  
# Plot statistics using only years with no missing data
plot_daily_stats(station_number = "08NM116",
                 complete_years = TRUE)

# Plot statistics for water years starting in October
plot_daily_stats(station_number = "08NM116",
                 start_year = 1980,
                 end_year = 2010,
                 water_year_start = 10)
                 
# Plot statistics with custom years
plot_daily_stats(station_number = "08NM116",
                 start_year = 1981,
                 end_year = 2010,
                 exclude_years = c(1991,1993:1995))
                  
# Plot statistics for 7-day flows for July-September months only
plot_daily_stats(station_number = "08NM116",
                 start_year = 1980,
                 roll_days = 7,
                 months = 7:9)

# Plot statistics without a log-scale Discharge axis
plot_daily_stats(station_number = "08NM116",
                 start_year = 1981,
                 end_year = 2010,
                 log_discharge = FALSE)
}
}
\seealso{
\code{\link{calc_daily_stats}}
}
