% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_screening.R
\name{plot_data_screening}
\alias{plot_data_screening}
\title{Plot annual summary statistics for data screening}
\usage{
plot_data_screening(data = NULL, dates = Date, values = Value,
  station_number = NULL, roll_days = 1, roll_align = "right",
  water_year = FALSE, water_year_start = 10, start_year = 0,
  end_year = 9999)
}
\arguments{
\item{data}{A data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
Leave blank if using \code{station_number} argument.}

\item{dates}{Column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not '
Date' (default). Leave blank if using \code{station_number} argument.}

\item{values}{Column in \code{data} that contains numeric flow values, in units of cubic metres per second.
Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.}

\item{station_number}{A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of
which to extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.
Leave blank if using \code{data} argument.}

\item{roll_days}{Numeric value of the number of days to apply a rolling mean. Default \code{1}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first 
('left'), last ('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the calendar year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}
}
\value{
A ggplot2 object with the following plots:
  \item{Minimum}{annual minimum of all daily flows for a given year}
  \item{Maximum}{annual maximum of all daily flows for a given year}
  \item{Mean}{annual mean of all daily flows for a given year}
  \item{StandardDeviation}{annual 1 standard deviation of all daily flows for a given year}
}
\description{
Plots the mean, median, maximum, minimum, standard deviation of annual flows. Plots the statistics from all daily 
   discharge values from all years, unless specified. Data calculated using screen_flow_data() function.
}
\examples{
\dontrun{

plot_data_screening(station_number = "08NM116", 
                    water_year = TRUE, 
                    water_year_start = 8)

}
}
