% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flow_data.R
\name{plot_flow_data}
\alias{plot_flow_data}
\title{Plot daily mean streamflow}
\usage{
plot_flow_data(data = NULL, dates = Date, values = Value,
  groups = STATION_NUMBER, rolling_days = 1, rolling_align = "right",
  water_year = FALSE, water_year_start = 10, start_year = 0,
  end_year = 9999, exclude_years = NULL, start_date = "0000-01-01",
  end_date = "3000-12-31", log_discharge = FALSE, plot_by_year = FALSE)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates, values, and (optional) groups (ex. station 
   names/numbers).
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{groups}{Column in the \code{data} data frame that contains unique identifiers for different data sets. 
Only required if using the data frame option of \code{data} and groups column is not named 'STATION_NUMBER'.
Function will automatically group by a column named 'STATION_NUMBER' if present. Remove the 'STATION_NUMBER' column or identify 
another non-existing column name to remove this grouping. Identify another column if desired. Default \code{STATION_NUMBER}.}

\item{rolling_days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{rolling_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{start_date}{Date (YYYY-MM-DD) of first date to consider for plotting. Leave blank if all years are required.}

\item{end_date}{Date (YYYY-MM-DD) of last date to consider for plotting. Leave blank if all years are required.}

\item{log_discharge}{Logical value to indicate plotting the discharge axis (Y-axis) on a logarithmic scale. Default \code{TRUE}.}

\item{plot_by_year}{Logical value to indicate whether to plot each year of data individually. Default \code{FALSE}.}
}
\value{
A ggplot2 object of daily flows from flow_data or HYDAT flow data provided
}
\description{
Plot the daily mean flow values from a streamflow dataset. Plots the statistics from all daily discharge values from all 
   years, unless specified. Can choose specific dates to start and end plotting. Can choose to plot out each year separately. Data 
   calculated using calc_daily_stats() function. Multiple groups/stations can be plotted if provided with the \code{groups} argument.
}
\examples{
\dontrun{

plot_flow_data(data = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
