% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flow_duration.R
\name{plot_flow_duration}
\alias{plot_flow_duration}
\title{Plot flow duration curves}
\usage{
plot_flow_duration(data = NULL, dates = Date, values = Value,
  roll_days = 1, roll_align = "right", water_year = FALSE,
  water_year_start = 10, start_year = 0, end_year = 9999,
  exclude_years = NULL, complete_years = FALSE, custom_months = NULL,
  custom_months_label = "Custom-Months", ignore_missing = TRUE,
  incl_months = 1:12, incl_longterm = TRUE, log_discharge = TRUE)
}
\arguments{
\item{data}{Daily data to be analyzed. Options:

   A data frame of daily data that contains columns of dates and values.
   
   A character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
   extract daily streamflow data from a HYDAT database. Requires \code{tidyhydat} package and a HYDAT database.}

\item{dates}{Column in the \code{data} data frame that contains dates formatted YYYY-MM-DD. Only required if
using the data frame option of \code{data} and dates column is not named 'Date'. Default \code{Date}.}

\item{values}{Column in the \code{data} data frame that contains numeric flow values, in units of cubic metres per second.
Only required if using the data frame option of \code{data} and values column is not named 'Value'. Default \code{Value}.}

\item{roll_days}{Numeric vector of the number of days to apply the rolling mean. Default \code{c(3,7,30)}.}

\item{roll_align}{Character string identifying the direction of the rolling mean from the specified date, either by the first ('left'), last
('right), or middle ('center') day of the rolling n-day group of observations. Default \code{'right'}.}

\item{water_year}{Logical value indicating whether to use water years to group data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Numeric value indicating the month of the start of the water year. Used if \code{water_year = TRUE}. 
Default \code{10}.}

\item{start_year}{Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.}

\item{end_year}{Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.}

\item{exclude_years}{Numeric vector of years to exclude from analysis. Leave blank to include all years.}

\item{complete_years}{Logical values indicating whether to include only years with complete data in analysis. Default \code{FALSE}.}

\item{custom_months}{Numeric vector of months to combine to summarize (ex. \code{6:8} for Jun-Aug). Adds results to the end of table.
If wanting months that overlap calendar years (ex. Oct-Mar), choose water_year and a water_year_month that begins before the first 
month listed. Leave blank for no custom month summary.}

\item{custom_months_label}{Character string to use as a label of custom months. For example, if choosing months 7:9  you may choose 
"Summer" or "Jul-Sep". Default \code{"Custom-Months"}.}

\item{ignore_missing}{Logical value indicating whether dates with missing values should be included in the calculation. If
\code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only statistics from time periods 
with no missing dates will be returned. Default \code{TRUE}.}

\item{incl_months}{Numeric vector of month curvesto plot. NA if no months required. Default \code{1:12}.}

\item{incl_longterm}{Logical value indicating whether to include longterm curve of all data. Default \code{TRUE}.}

\item{log_discharge}{Logical value to indicate plotting the discharge axis (Y-axis) on a logarithmic scale. Default \code{TRUE}.}
}
\value{
A ggplot2 object with plots for each month, long-term, and custom months showing the percentage of time that 
   flows are likely equal or exceeded for each time period.
}
\description{
Plots flow duration curves, percent time a flow value is equalled or exceeded, for a streamflow dataset. Plots 
   statistics from all daily discharge values from all years, unless specified. Data calculated using calc_longterm_stats() 
   function then converted for plotting.
}
\examples{
\dontrun{

plot_flow_duration(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
