% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_longterm_stats.R
\name{plot_longterm_stats}
\alias{plot_longterm_stats}
\title{Plot the long-term and long-term monthly summary statistics}
\usage{
plot_longterm_stats(flow_data = NULL, flow_dates = Date,
  flow_values = Value, flow_stations = STATION_NUMBER, HYDAT = NULL,
  plot_station = NULL, water_year = FALSE, water_year_start = 10,
  start_year = 0, end_year = 9999, exclude_years = NULL,
  log_discharge = TRUE)
}
\arguments{
\item{flow_data}{a data frame of daily mean flow data that contains columns of dates, flow values, and (optional) station 
names/numbers. Leave blank if using \code{HYDAT} argument.}

\item{flow_dates}{a column in flow_data that contains dates of daily flow data formatted YYYY-MM-DD. Leave blank if using \code{HYDAT} 
argument. Default \code{Date}.}

\item{flow_values}{a column in flow_data that contains numeric values of daily mean flow data, in units of cubic metres per second. 
Leave blank if using \code{HYDAT} argument. Default \code{Value}.}

\item{flow_stations}{a column in flow_data that contains station identifiers for each flow data set.
Removing 'STATION_NUMBER' column in flow_data or incorrectly identifying will calculate statistics on all flow values from all stations.
If using \code{HYDAT} argument, setting \code{flow_stations} to anything besides \code{STATION_NUMBER} will have similar effects.
Default \code{STATION_NUMBER}. Must select just one station to plot using \code{plot_station} argument.}

\item{HYDAT}{a character string vector of seven digit Water Survey of Canada station numbers (e.g. \code{"08NM116"}) of which to 
extract daily streamflow data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the 
\code{tidyhydat} package and a HYDAT database are required. Leave blank if using \code{flow_data} arguments.}

\item{plot_station}{a character string identifying the station listed in the flow data/HYDAT data to plot if multiple stations are 
provided.}

\item{water_year}{a logical value indicating whether to use water years to group flow data instead of calendar years. Water years 
are designated by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{a numeric value indicating the month of the start of the water year. Used if \code{water_year=TRUE}. 
Default \code{10}.}

\item{start_year}{a numeric value of the first year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{a numeric value of the last year to consider for analysis. Leave blank if all years are required.}

\item{exclude_years}{a numeric vector of years to exclude from analysis. Leave blank if all years are required.}

\item{log_discharge}{a logical value to indicate plotting thedischarge axis (Y-axis) on a logarithmic scale. Default \code{TRUE}.}
}
\value{
A ggplot2 object with the following plots:
  \item{Long-term Mean}{mean of all daily flows over all years}
  \item{Long-term Median}{median of all daily flows over all years}
  \item{Monthly Mean}{mean of all daily flows for each month over all years}
  \item{Monthly Median}{median of all daily flows for each month over all years}
  \item{25-75 Percentiles Range}{a ribbon showing the range of data between the monthly 25th and 75th percentiles}
  \item{5-95 Percentiles Range}{a ribbon showing the range of data between the monthly 5th and 95th percentiles}
  \item{Max-Min Range}{a ribbon showing the range of data between the monthly minimum and maximums}
}
\description{
Plots the long-term and long-term monthly mean, median, maximum, minimum, and 5, 25, 75, and 95 percentiles of daily 
   flow values from a single streamflow dataset. Plots statistics from all daily discharge values from all years, unless specified. 
   Data calculated using calc_longterm_stats() function.
}
\examples{
\dontrun{

plot_longterm_stats(flow_data = flow_data)

plot_longterm_stats(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
