% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_flow_data.R
\name{screen_flow_data}
\alias{screen_flow_data}
\title{Calculate annual summary and missing data statistics}
\usage{
screen_flow_data(flowdata = NULL, HYDAT = NULL, rolling_days = 1,
  rolling_align = "right", water_year = FALSE, water_year_start = 10,
  start_year = NULL, end_year = NULL, transpose = FALSE,
  station_name = NA, write_table = FALSE, write_digits = 3,
  write_dir = ".")
}
\arguments{
\item{flowdata}{Data frame. A data frame of daily mean flow data that includes two columns: a 'Date' column with dates formatted 
YYYY-MM-DD, and a numeric 'Value' column with the corresponding daily mean flow values in units of cubic metres per second. 
Not required if \code{HYDAT} argument is used.}

\item{HYDAT}{Character. A seven digit Water Survey of Canada station number (e.g. \code{"08NM116"}) of which to extract daily streamflow 
data from a HYDAT database. \href{https://github.com/ropensci/tidyhydat}{Installation} of the \code{tidyhydat} package and a HYDAT 
database are required. Not required if \code{flowdata} argument is used.}

\item{rolling_days}{Numeric. The number of days to apply a rolling mean. Default \code{1}.}

\item{rolling_align}{Character. Specifies whether the dates of the rolling mean should be specified by the first ('left'), last ('right),
or middle ('center') of the rolling n-day group of observations. Default \code{'right'}.
by the year in which they end. Default \code{FALSE}.}

\item{water_year}{Logical. Use water years to group flow data instead of calendar years. Water years are designated
by the year in which they end. Default \code{FALSE}.}

\item{water_year_start}{Integer. Month indicating the start of the water year. Used if \code{water_year=TRUE}. Default \code{10}.}

\item{start_year}{Integer. First year to consider for analysis. Leave blank if all years are required.}

\item{end_year}{Integer. Last year to consider for analysis. Leave blank if all years are required.}

\item{transpose}{Logical. Switch the rows and columns of the results table. Default \code{FALSE}.}

\item{station_name}{Character. Name of hydrometric station or stream that will be used to create file names. Leave blank if not writing
files or if \code{HYDAT} is used or a column in \code{flowdata} called 'STATION_NUMBER' contains a WSC station number, as the name
will be the \code{HYDAT} value provided in the argument or column. Setting the station name will replace the HYDAT station number.}

\item{write_table}{Logical. Write the table as a .csv file to specified directory. Default \code{FALSE}.}

\item{write_digits}{Numeric. Number of significant digits to round the results in the written table. Default \code{3}.}

\item{write_dir}{Character. Directory folder name of where to write tables and plots. If directory does not exist, it will be created.
Default is the working directory.}
}
\value{
A data frame with the following columns:
  \item{Year}{calendar or water year selected}
  \item{n_days}{number of days per year}
  \item{n_Q}{number of days per year with flow data}
  \item{n_missing_Q}{number of days per year with no flow data}
  \item{Minimum}{annual minimum of all daily flows for a given year}
  \item{Maximum}{annual maximum of all daily flows for a given year}
  \item{Mean}{annual mean of all daily flows for a given year}
  \item{Median}{annual median of all daily flows for a given year}
  \item{StandardDeviation}{annual 1 standard deviation of all daily flows for a given year}
  and the following monthly missing columns (order will depend on water_year_month):
  \item{Jan_missing_Q}{number of Jan days per year with no flow data}
  \item{Feb_missing_Q}{number of Feb days per year with no flow data}
  \item{Mar_missing_Q}{number of Mar days per year with no flow data}
  \item{Apr_missing_Q}{number of Apr days per year with no flow data}
  \item{May_missing_Q}{number of May days per year with no flow data}
  \item{Jun_missing_Q}{number of Jun days per year with no flow data}
  \item{Jul_missing_Q}{number of Jul days per year with no flow data}
  \item{Aug_missing_Q}{number of Aug days per year with no flow data}
  \item{Sep_missing_Q}{number of Sep days per year with no flow data}
  \item{Oct_missing_Q}{number of Oct days per year with no flow data}
  \item{Nov_missing_Q}{number of Nov days per year with no flow data}
  \item{Dec_missing_Q}{number of Dec days per year with no flow data}
  Transposing data creates a column of "Statistics" and subsequent columns for each year selected.
}
\description{
Calculates mean, median, maximum, minimum, standard deviation of annual flows and data availability and missing data
   statistics for each year and month of each year. Calculates the statistics from all daily discharge values from all years, 
   unless specified.
}
\examples{
\dontrun{

screen_flow_data(flowdata = flowdata, station_name = "MissionCreek", write_table = TRUE)

screen_flow_data(HYDAT = "08NM116", water_year = TRUE, water_year_start = 8)

}
}
