% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plots.R
\name{write_plots}
\alias{write_plots}
\title{Write a plot in a directory}
\usage{
write_plots(plot = NULL, file = "", format = NULL, height = NA,
  width = NA, units = "in", combined_pdf = FALSE)
}
\arguments{
\item{plot}{Name of plot or name of list of plots to write to a directory.}

\item{file}{Character string of name of file (single or combined_pdf) or name of folder (mulitple plots) in a directory 
to write the plot(s). Listing the image format (ex. ".png" or ".pdf") will overwrite the \code{format} argument for
single plots.}

\item{format}{Character string of image type to write. One of "png", "eps", "ps", "tex", "pdf", "jpeg", "tiff", "bmp", or "svg".
Image type will be overwritten if using \code{combined_pdf argument}.}

\item{height}{Numeric plot height in \code{units}. If not supplied, uses the size of current graphics device.}

\item{width}{Numeric plot width in \code{units}. If not supplied, uses the size of current graphics device.}

\item{units}{Character string plot height and width units, one of "in", "cm", or "mm". Default \code{"in"}.}

\item{combined_pdf}{Logical value indicating whether to combine list of plots into one pdf document. Default \code{FALSE}.}
}
\description{
Write plots, single or multiple, into a folder or pdf document
}
\examples{
\dontrun{

write_plots()

}
}
